% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-argument-glossary.R
\docType{data}
\name{argument_glossary}
\alias{argument_glossary}
\title{Allowed argument names in tidiers}
\format{A tibble with 3 variables:
\describe{
\item{method}{One of "glance", "augment" or "tidy".}
\item{argument}{Character name of allowed argument name.}
\item{description}{Character description of argument use.}
}}
\usage{
argument_glossary
}
\description{
Allowed argument names in tidiers
}
\examples{
argument_glossary
}
\keyword{datasets}
