% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_rw_tibble}
\alias{as_rw_tibble}
\title{Coerce a data frame to a tibble, preserving rownames}
\usage{
as_rw_tibble(data)
}
\arguments{
\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \code{tibble} potentially with a \code{.rownames} column
}
\description{
A thin wrapper around \code{\link[tibble:as_tibble]{tibble::as_tibble()}}, except checks for
rownames and adds them to a new column \code{.rownames} if they are
interesting (i.e. more than \code{1, 2, 3, ...}).
}
\details{
Replacement for \code{fix_data_frame()}.
}
