% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg-tidiers.R
\name{augment.betareg}
\alias{augment.betareg}
\title{Augment data with information from a(n) betareg object}
\usage{
\method{augment}{betareg}(x, data = stats::model.frame(x),
  newdata = NULL, type.predict, type.residuals, ...)
}
\arguments{
\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors usednto create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{type.predict}{Character indicating type of prediction to use. TODO:
Figure out what the options are, how they are different, and make
sure to use arg.match here.}

\item{type.residuals}{Character indicating type of residuals to use. TODO:
Figure out what the options are, how they are different, and make
sure to use arg.match here.}
}
\value{
augment returns the original data, along with new columns describing
each observation:
\item{.fitted}{Fitted values of model}
\item{.resid}{Residuals}
\item{.cooksd}{Cooks distance, \code{\link[=cooks.distance]{cooks.distance()}}}
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[betareg:betareg]{betareg::betareg()}}
}
