% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-glmrob-tidiers.R
\name{augment.glmRob}
\alias{augment.glmRob}
\title{Tidy a(n) glmRob object}
\usage{
\method{augment}{glmRob}(x, ...)
}
\arguments{
\item{x}{A \code{glmRob} object returned from \code{\link[robust:glmRob]{robust::glmRob()}}.}

\item{...}{Arguments passed on to \code{augment.lm}
\describe{
  \item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}
  \item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors usednto create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}
  \item{type.predict}{Type of prediction to compute for a GLM; passed on to
\code{\link[=predict.glm]{predict.glm()}}}
  \item{type.residuals}{Type of residuals to compute for a GLM; passed on to
\code{\link[=residuals.glm]{residuals.glm()}}}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{augment} method does goes here. Wraps
\code{\link[=augment.lm]{augment.lm()}}. Details on this. Document params other than \code{x}.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[=augment.lm]{augment.lm()}}

\code{\link[robust:glmRob]{robust::glmRob()}}

Other robust tidiers: \code{\link{augment.lmRob}},
  \code{\link{glance.glmRob}}, \code{\link{glance.lmRob}},
  \code{\link{tidy.glmRob}}, \code{\link{tidy.lmRob}}
}
\concept{robust tidiers}
