% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-htest-tidiers.R
\name{augment.htest}
\alias{augment.htest}
\title{Augment data with information from a(n) htest object}
\usage{
\method{augment}{htest}(x, ...)
}
\arguments{
\item{x}{An \code{htest} objected, such as those created by \code{\link[stats:cor.test]{stats::cor.test()}},
\code{\link[stats:t.test]{stats::t.test()}}, \code{\link[stats:wilcox.test]{stats::wilcox.test()}}, \code{\link[stats:chisq.test]{stats::chisq.test()}}, etc.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
Errors unless \code{x} is a chi-squared test. If \code{x} is a chi-squared
test, for each cell of the tested table returns columns:

\item{.observed}{Observed count}
\item{.prop}{Proportion of the total}
\item{.row.prop}{Row proportion (2 dimensions table only)}
\item{.col.prop}{Column proportion (2 dimensions table only)}
\item{.expected}{Expected count under the null hypothesis}
\item{.residuals}{Pearson residual}
\item{.stdres}{Standardized residual}
}
\description{
Description of what a \code{augment} method does goes here.
}
\details{
See \code{\link[stats:chisq.test]{stats::chisq.test()}} for more details on
how residuals are computed.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[stats:chisq.test]{stats::chisq.test()}}

Other htest tidiers: \code{\link{tidy.htest}},
  \code{\link{tidy.pairwise.htest}},
  \code{\link{tidy.power.htest}}
}
\concept{htest tidiers}
