% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-lm-tidiers.R
\name{augment.lm}
\alias{augment.lm}
\title{Augment data with information from a(n) lm object}
\usage{
\method{augment}{lm}(x, data = stats::model.frame(x), newdata,
  type.predict, type.residuals, ...)
}
\arguments{
\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors usednto create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{type.predict}{Type of prediction to compute for a GLM; passed on to
\code{\link[=predict.glm]{predict.glm()}}}

\item{type.residuals}{Type of residuals to compute for a GLM; passed on to
\code{\link[=residuals.glm]{residuals.glm()}}}
}
\value{
When \code{newdata} is not supplied \code{augment.lm} returns
one row for each observation, with seven columns added to the original
data:

\item{.hat}{Diagonal of the hat matrix}
\item{.sigma}{Estimate of residual standard deviation when
corresponding observation is dropped from model}
\item{.cooksd}{Cooks distance, \code{\link[=cooks.distance]{cooks.distance()}}}
\item{.fitted}{Fitted values of model}
\item{.se.fit}{Standard errors of fitted values}
\item{.resid}{Residuals}
\item{.std.resid}{Standardised residuals}

(Some unusual "lm" objects, such as "rlm" from MASS, may omit
\code{.cooksd} and \code{.std.resid}. "gam" from mgcv omits
\code{.sigma})

When \code{newdata} is supplied, \code{augment.lm} returns one row for each
observation, with three columns added to the new data:

\item{.fitted}{Fitted values of model}
\item{.se.fit}{Standard errors of fitted values}
\item{.resid}{Residuals of fitted values on the new data}
}
\description{
Description of what a \code{augment} method does goes here.
}
\details{
When the modeling was performed with \code{na.action = "na.omit"}
(as is the typical default), rows with NA in the initial data are omitted
entirely from the augmented data frame. When the modeling was performed
with \code{na.action = "na.exclude"}, one should provide the original data
as a second argument, at which point the augmented data will contain those
rows (typically with NAs in place of the new columns). If the original data
is not provided to \code{augment} and \code{na.action = "na.exclude"}, a
warning is raised and the incomplete rows are dropped.
}
\seealso{
\link{na.action}

\code{\link[=augment]{augment()}}, \code{\link[stats:predict.lm]{stats::predict.lm()}}

Other lm tiders: \code{\link{glance_lm}},
  \code{\link{lm_tidiers}}
}
\concept{lm tiders}
