% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loess_tidiers.R
\name{augment.loess}
\alias{augment.loess}
\title{Augmenting methods for loess models}
\usage{
\method{augment}{loess}(x, data = model.frame(x), se = TRUE, ...)
}
\value{
When \code{newdata} is not supplied \code{augment.loess}
returns one row for each observation with three columns added
to the original data (two if \code{se = FALSE}):
   \item{.fitted}{Fitted values of model}
   \item{.se.fit}{Standard errors of the fitted values}
   \item{.resid}{Residuals of the fitted values}

   When \code{newdata} is supplied \code{augment.loess} returns
   one row for each observation with one additional column (two
   if \code{se = TRUE}):
   \item{.fitted}{Fitted values of model}
   \item{.se.fit}{Standard errors of the fitted values}
}
\description{
This method augments the original data with information
on the fitted values and residuals, and optionally the
standard errors.
}

