% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust-tidiers.R
\name{augment.Mclust}
\alias{augment.Mclust}
\title{Augment data with information from a(n) Mclust object}
\usage{
\method{augment}{Mclust}(x, data, ...)
}
\arguments{
\item{x}{An \code{Mclust} object return from \code{\link[mclust:Mclust]{mclust::Mclust()}}.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of the original data with two extra columns:
\item{.class}{The class assigned by the Mclust algorithm}
\item{.uncertainty}{The uncertainty associated with the classification.
If a point has a probability of 0.9 of being in its assigned class
under the model, then the uncertainty is 0.1.}
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[mclust:Mclust]{mclust::Mclust()}}

Other mclust tidiers: \code{\link{tidy.Mclust}}
}
