% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfx-tidiers.R
\name{augment.mfx}
\alias{augment.mfx}
\alias{mfx_tidiers}
\title{Augment data with information from a(n) mfx object}
\usage{
\method{augment}{mfx}(x, data = model.frame(x$fit), newdata = NULL,
  type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)
}
\arguments{
\item{x}{A \code{betamfx}, \code{logitmfx}, \code{negbinmfx}, \code{poissonmfx}, or \code{probitmfx}  object.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors used to create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{type.predict}{Passed to \code{\link[stats:predict.glm]{stats::predict.glm()}} \code{type}
argument. Defaults to \code{"link"}.}

\item{type.residuals}{Passed to \code{\link[stats:residuals.glm]{stats::residuals.glm()}} and
to \code{\link[stats:rstandard.glm]{stats::rstandard.glm()}} \code{type} arguments. Defaults to \code{"deviance"}.}

\item{se_fit}{Logical indicating whether or not a \code{.se.fit} column should be
added to the augmented output. For some models, this calculation can be
somwhat time-consuming. Defaults to \code{FALSE}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Augment accepts a model object and a dataset and adds
information about each observation in the dataset. Most commonly, this
includes predicted values in the \code{.fitted} column, residuals in the
\code{.resid} column, and standard errors for the fitted values in a \code{.se.fit}
column. New columns always begin with a \code{.} prefix to avoid overwriting
columns in the original dataset.

Users may pass data to augment via either the \code{data} argument or the
\code{newdata} argument. If the user passes data to the \code{data} argument,
it \strong{must} be exactly the data that was used to fit the model
object. Pass datasets to \code{newdata} to augment data that was not used
during model fitting. This still requires that all columns used to fit
the model are present.

Augment will often behavior different depending on whether \code{data} or
\code{newdata} is specified. This is because there is often information
associated with training observations (such as influences or related)
measures that is not meaningfully defined for new observations.

For convenience, many augment methods provide default \code{data} arguments,
so that \code{augment(fit)} will return the augmented training data. In these
cases augment tries to reconstruct the original data based on the model
object, with some varying degrees of success.

The augmented dataset is always returned as a \link[tibble:tibble]{tibble::tibble} with the
\strong{same number of rows} as the passed dataset. This means that the
passed data must be coercible to a tibble. At this time, tibbles do not
support matrix-columns. This means you should not specify a matrix
of covariates in a model formula during the original model fitting
process, and that \code{\link[splines:ns]{splines::ns()}}, \code{\link[stats:poly]{stats::poly()}} and
\code{\link[survival:Surv]{survival::Surv()}} objects are not supported in input data. If you
encounter errors, try explicitly passing a tibble, or fitting the original
model on data in a tibble.

We are in the process of defining behaviors for models fit with various
\link{na.action} arguments, but make no guarantees about behavior when data is
missing at this time.
}
\details{
This generic augment method wraps \code{\link[=augment.glm]{augment.glm()}} for objects from
the \code{mfx} package.
}
\examples{
library(mfx)
mfx_logitreg <- logitmfx(am ~ cyl + hp + wt, atmean = T, data = mtcars)
augment(mfx_logitreg)
mfx_probitreg <- probitmfx(am ~ cyl + hp + wt, atmean = T, data = mtcars)
augment(mfx_probitreg)
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[=augment.glm]{augment.glm()}}, \code{\link[mfx:betamfx]{mfx::betamfx()}}, \code{\link[mfx:logitmfx]{mfx::logitmfx()}}, \code{\link[mfx:negbinmfx]{mfx::negbinmfx()}}, \code{\link[mfx:poissonmfx]{mfx::poissonmfx()}}, \code{\link[mfx:probitmfx]{mfx::probitmfx()}}

Other mfx tidiers: \code{\link{glance.mfx}}
}
\concept{mfx tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{.cooksd}{Cooks distance.}
  \item{.fitted}{Fitted or predicted value.}
  \item{.hat}{Diagonal of the hat matrix.}
  \item{.resid}{The difference between fitted and observed values.}
  \item{.se.fit}{Standard errors of fitted values.}
  \item{.sigma}{Estimated residual standard deviation when corresponding observation is dropped from model.}
  \item{.std.resid}{Standardised residuals.}

}
