% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinerml-tidiers.R
\name{augment.mjoint}
\alias{augment.mjoint}
\title{Augment data with information from a(n) mjoint object}
\usage{
\method{augment}{mjoint}(x, data = x$data, ...)
}
\arguments{
\item{x}{An \code{mjoint} object returned from \code{\link[joineRML:mjoint]{joineRML::mjoint()}}.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each original observation
with addition columns:
\item{.fitted_j_0}{population-level fitted values for the
j-th longitudinal process}
\item{.fitted_j_1}{individuals-level fitted values for the j-th
longitudinal process}
\item{.resid_j_0}{population-level residuals for the j-th
longitudinal process}
\item{.resid_j_1}{individual-level residuals for the j-th
longitudinal process}
}
\description{
Description of what a \code{augment} method does goes here.
}
\details{
See \code{\link[joineRML:fitted.mjoint]{joineRML::fitted.mjoint()}} and \code{\link[joineRML:residuals.mjoint]{joineRML::residuals.mjoint()}} for
more information on the difference between population-level and
individual-level fitted values and residuals.

If fitting a joint model with a single longitudinal process,
make sure you are using a named \code{list} to define the formula
for the fixed and random effects of the longitudinal submodel.
}
