% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-mvr-tidiers.R
\name{augment.mvr}
\alias{augment.mvr}
\title{Augment data with information from a(n) mvr object}
\usage{
\method{augment}{mvr}(x, data = model.frame(x), newdata = NULL,
  ncomp = x$ncomp, ...)
}
\arguments{
\item{x}{a mvr object such as those created by \code{\link[pls:mvr]{pls::mvr()}}, \code{\link[pls:plsr]{pls::plsr()}}, \code{\link[pls:cppls]{pls::cppls()}} and
\code{\link[pls:pcr]{pls::pcr()}}.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors used to create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{ncomp}{The number of components to include in the model. Ignored if comps is specified.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Augment accepts a model object and a dataset and adds
information about each observation in the dataset. Most commonly, this
includes predicted values in the \code{.fitted} column, residuals in the
\code{.resid} column, and standard errors for the fitted values in a \code{.se.fit}
column. New columns always begin with a \code{.} prefix to avoid overwriting
columns in the original dataset.

Users may pass data to augment via either the \code{data} argument or the
\code{newdata} argument. If the user passes data to the \code{data} argument,
it \strong{must} be exactly the data that was used to fit the model
object. Pass datasets to \code{newdata} to augment data that was not used
during model fitting. This still requires that all columns used to fit
the model are present.

Augment will often behavior different depending on whether \code{data} or
\code{newdata} is specified. This is because there is often information
associated with training observations (such as influences or related)
measures that is not meaningfully defined for new observations.

For convenience, many augment methods provide default \code{data} arguments,
so that \code{augment(fit)} will return the augmented training data. In these
cases augment tries to reconstruct the original data based on the model
object, with some varying degrees of success.

The augmented dataset is always returned as a \link[tibble:tibble]{tibble::tibble} with the
\strong{same number of rows} as the passed dataset. This means that the
passed data must be coercible to a tibble. At this time, tibbles do not
support matrix-columns. This means you should not specify a matrix
of covariates in a model formula during the original model fitting
process, and that \code{\link[splines:ns]{splines::ns()}}, \code{\link[stats:poly]{stats::poly()}} and
\code{\link[survival:Surv]{survival::Surv()}} objects are not supported in input data. If you
encounter errors, try explicitly passing a tibble, or fitting the original
model on data in a tibble.

We are in the process of defining behaviors for models fit with various
\link{na.action} arguments, but make no guarantees about behavior when data is
missing at this time.
}
\examples{
library(pls)
library(dplyr)
library(ggplot2)

data(yarn) # Single-response model
yarn.pls <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
glance(yarn.pls)
tidy(yarn.pls)
augment(yarn.pls)

data(oliveoil) # Multi-response model
sens.pls <- plsr(sensory ~ chemical, ncomp = 4, scale = TRUE, data = oliveoil)
glance(sens.pls)
tidy(sens.pls)
augment(sens.pls)

set.seed(42)
test_id <- sample(1:nrow(yarn), nrow(yarn) / 3)
yarn.pls.train <- plsr(density ~ NIR, 3, data = slice(yarn, -test_id))
augment(yarn.pls.train, newdata = slice(yarn, test_id)) \%>\%
  mutate(.rownames = test_id) \%>\%
  ggplot(aes(.rownames, .t.squared[, 3])) +
  geom_point() +
  ggtitle("How far each sample is from the center of the model.")
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[pls:predict.mvr]{pls::predict.mvr()}}

Other mvr tidiers: \code{\link{glance.mvr}},
  \code{\link{tidy.mvr}}
}
\concept{mvr tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{response}{Which response column the coefficients correspond to (typically `Y1`, `Y2`, etc).}
  \item{.t.squared}{Hotelling's T^2 value.}

}
