% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg-nlrq-tidiers.R
\name{augment.nlrq}
\alias{augment.nlrq}
\title{Tidy a(n) nlrq object}
\usage{
\method{augment}{nlrq}(x, ...)
}
\arguments{
\item{x}{A \code{nlrq} object returned from \code{\link[quantreg:nlrq]{quantreg::nlrq()}}.}

\item{...}{Arguments passed on to \code{augment.nls}
\describe{
  \item{x}{An object of class "nls"}
  \item{data}{original data this was fitted on; if not given this will
attempt to be reconstructed from nls (may not be successful)}
  \item{newdata}{new data frame to use for predictions}
}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the data passed to \code{augment},
along with additional columns:
\item{.fitted}{The predicted response for that observation.}
\item{.resid}{The residual for a particular point. Present only when
data has been passed to \code{augment} via the \code{data} argument.}
}
\description{
Description of what a \code{tidy} method does goes here.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[quantreg:nlrq]{quantreg::nlrq()}}

Other quantreg tidiers: \code{\link{augment.rqs}},
  \code{\link{augment.rq}}, \code{\link{glance.nlrq}},
  \code{\link{glance.rq}}, \code{\link{tidy.nlrq}},
  \code{\link{tidy.rqs}}, \code{\link{tidy.rq}}
}
\concept{quantreg tidiers}
