% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-nls-tidiers.R
\name{augment.nls}
\alias{augment.nls}
\title{Augment data with information from a(n) nls object}
\usage{
\method{augment}{nls}(x, data = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{An \code{nls} object returned from \code{\link[stats:nls]{stats::nls()}}.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors usednto create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the data passed to \code{augment},
along with additional columns:
\item{.fitted}{The predicted response for that observation.}
\item{.resid}{The residual for a particular point. Present only when
data has been passed to \code{augment} via the \code{data} argument.}
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\link{tidy}, \code{\link[stats:nls]{stats::nls()}}, \code{\link[stats:predict.nls]{stats::predict.nls()}}

Other nls tidiers: \code{\link{glance.nls}},
  \code{\link{tidy.nls}}
}
\concept{nls tidiers}
