% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polca-tidiers.R
\name{augment.poLCA}
\alias{augment.poLCA}
\title{Augment data with information from a(n) poLCA object}
\usage{
\method{augment}{poLCA}(x, data = NULL, ...)
}
\arguments{
\item{x}{A \code{poLCA} object returned from \code{\link[poLCA:poLCA]{poLCA::poLCA()}}.}

\item{data}{The original dataset used to fit the latent class model, as
a tibble or data. If not given, uses manifest variables in
\code{x$y} and, if applicable, covariates in \code{x$x}}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row for each original observation,
with additional columns:

\item{.class}{Predicted class, using modal assignment}
\item{.probability}{Posterior probability of predicted class}

If the \code{data} argument is given, those columns are included in the output
(only rows for which predictions could be made).
Otherwise, the \code{y} element of the poLCA object, which contains the
manifest variables used to fit the model, are used, along with any covariates,
if present, in \code{x}.

Note that while the probability of all the classes (not just the predicted
modal class) can be found in the \code{posterior} element, these are not
included in the augmented output.
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[poLCA:poLCA]{poLCA::poLCA()}}

Other poLCA tidiers: \code{\link{glance.poLCA}},
  \code{\link{tidy.poLCA}}
}
\concept{poLCA tidiers}
