% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-prcomp-tidiers.R
\name{augment.prcomp}
\alias{augment.prcomp}
\title{Augment data with information from a(n) prcomp object}
\usage{
\method{augment}{prcomp}(x, data = NULL, newdata, ...)
}
\arguments{
\item{x}{A \code{prcomp} object returned by \code{\link[stats:prcomp]{stats::prcomp()}}.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors usednto create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing the original data along with
additional columns containing each observation's projection into
PCA space.
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[stats:prcomp]{stats::prcomp()}}, \link{svd_tidiers}

Other svd tidiers: \code{\link{tidy.prcomp}},
  \code{\link{tidy_irlba}}, \code{\link{tidy_svd}}
}
\concept{svd tidiers}
