% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{augment}
\alias{augment}
\title{Augment data according to a tidied model}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{Model object or other R object with information to append to
observations.}

\item{...}{Addition arguments to augment method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about data points.
}
\description{
Given an R statistical model or other non-tidy object, add columns to the
original dataset such as predictions, residuals and cluster assignments.
}
\details{
Note that by convention the first argument is almost always \code{data},
which specifies the original data object. This is not part of the S3
signature, partly because it prevents \link{rowwise_df_tidiers} from
taking a column name as the first argument.

This generic originated in the ggplot2 package, where it was called
\code{fortify}.
}
\seealso{
\code{\link[=augment.lm]{augment.lm()}}
}
