% Generated by roxygen2 (4.0.1): do not edit by hand
\name{augment.smooth.spline}
\alias{augment.smooth.spline}
\title{augment method for smooth.spline}
\usage{
\method{augment}{smooth.spline}(model, data = model$data, ...)
}
\arguments{
\item{model}{a smooth.spline object}

\item{data}{defaults to data used to fit model}

\item{...}{not used in this method}
}
\value{
The original data with extra columns:
  \item{.fitted}{Fitted values of model}
  \item{.resid}{Residuals}
}
\description{
This combines the original data given to smooth.spline with the
fit and residuals
}

