% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-decompose-tidiers.R
\name{augment.stl}
\alias{augment.stl}
\title{Augment data with information from a(n) stl object}
\usage{
\method{augment}{stl}(x, weights = TRUE, ...)
}
\arguments{
\item{x}{An \code{stl} object returned from \code{\link[stats:stl]{stats::stl()}}.}

\item{weights}{Logical indicating whether or not to include the robust
weights in the output.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one row for each observation in the
original times series:

\item{\code{.seasonal}}{The seasonal component of the decomposition.}
\item{\code{.trend}}{The trend component of the decomposition.}
\item{\code{.remainder}}{The remainder, or "random" component of the
decomposition.}
\item{\code{.weight}}{The final robust weights, if requested.}
\item{\code{.seasadj}}{The seasonally adjusted (or "deseasonalised")
series.}
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[stats:stl]{stats::stl()}}

Other decompose tidiers: \code{\link{augment.decomposed.ts}}
}
\concept{decompose tidiers}
