% Generated by roxygen2 (4.0.2): do not edit by hand
\name{augment_columns}
\alias{augment_columns}
\title{add predictions, fitted values, residuals, and other common outputs from
an augment call}
\usage{
augment_columns(x, newdata, type, type.predict = type,
  type.residuals = type, ...)
}
\arguments{
\item{x}{a model}

\item{newdata}{new data to predict on, optional}

\item{type}{Type of prediction and residuals to compute}

\item{type.predict}{Type of prediction to compute; by default
same as \code{type}}

\item{type.residuals}{Type of residuals to compute; by default
same as \code{type}}

\item{...}{extra arguments (not used)}
}
\description{
In the case that a generic is not implemented for the model, fail quietly.
}

