% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{augment.rma}
\alias{augment.rma}
\title{Augment data with values from a meta-analysis model}
\usage{
\method{augment}{rma}(x, ..., data = NULL)
}
\arguments{
\item{x}{a meta-analysis object. Currently supports \code{rma.uni} from the
\code{metafor} package.}

\item{...}{additional arguments}
}
\value{
a \code{data.frame}
}
\description{
Augment the original data with model residuals, fitted values, and influence
statistics.
}
\examples{

example_ma \%>\%
  augment()

}
