% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot_tidiers.R
\name{boot_tidiers}
\alias{boot_tidiers}
\alias{tidy.boot}
\title{Tidying methods for  bootstrap computations}
\usage{
\method{tidy}{boot}(x, conf.int = FALSE, conf.level = 0.95,
  conf.method = "perc", ...)
}
\arguments{
\item{x}{\code{\link{boot}} object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for CI}

\item{conf.method}{method for computing confidence intervals (see \code{\link{boot.ci}})}

\item{\dots}{extra arguments (not used)}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.
}
\description{
Tidying methods for  bootstrap computations
}
\examples{
if (require("boot")) {
}
if (require("boot")) {
   g1 <- glm(lot2 ~ log(u), data = clotting, family = Gamma)
   bootfun <- function(d, i) {
      coef(update(g1, data= d[i,]))
   }
   bootres <- boot(clotting, bootfun, R = 999)
   tidy(g1,conf.int=TRUE)
   tidy(bootres,conf.int=TRUE)
}
}

