% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_tidiers.R
\name{caret_tidiers}
\alias{caret_tidiers}
\alias{caret_tidiers}
\alias{tidy.confusionMatrix}
\title{Tidying methods for confusionMatrix objects}
\usage{
\method{tidy}{confusionMatrix}(x, show_class = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{confusionMatrix}}

\item{show_class}{A boolean of whether to show the values for class specific
quantities from the confusion matrix (specificty, sensitivity, etc.). If set
to FALSE, result will only show accuracy and kappa.}

\item{...}{extra arguments (not used)}
}
\description{
Tidies the result of confusion matrix from the \code{caret} package.
Only a \code{tidy} method is provided, not an \code{augment} or
\code{glance} method.
}
\examples{

# 2 class confusion matrix
cm2 <- caret::confusionMatrix(factor(rbinom(100,1,.5)),factor(rbinom(100,1,.5)))
tidy(cm2)
tidy(cm2, show_class = FALSE) # only shows accuracy and kappa

# confusion matrix with more than 2 classes
cm <- caret::confusionMatrix(factor(rbinom(100,2,.5)),factor(rbinom(100,2,.5)))
tidy(cm)
tidy(cm, show_class = FALSE) # only shows accuracy and kappa

}
