% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-column-glossary.R
\docType{data}
\name{column_glossary}
\alias{column_glossary}
\title{Allowed column names in tidied tibbles}
\format{A tibble with 4 variables:
\describe{
\item{method}{One of "glance", "augment" or "tidy".}
\item{column}{Character name of allowed output column.}
\item{description}{Character description of expected column contents.}
\item{used_by}{A list of character vectors detailing the \emph{classes}
that use the column when tidied. For example \code{c("Arima", "betareg")}.}
}}
\usage{
column_glossary
}
\description{
Allowed column names in tidied tibbles
}
\examples{
column_glossary
}
\keyword{datasets}
