% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_tidiers.R
\name{confint.geeglm}
\alias{confint.geeglm}
\title{Confidence interval for \code{geeglm} objects}
\usage{
\method{confint}{geeglm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{The 'geeglm' object}

\item{parm}{The parameter to calculate the confidence interval
for.  If not specified, the default is to calculate a confidence
interval on all parameters (all variables in the model).}

\item{level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{...}{Additional parameters}
}
\value{
Returns the upper and lower confidence intervals
}
\description{
Generate confidence intervals for GEE analyses
}
\details{
This function was taken from
http://stackoverflow.com/a/21221995/2632184.
}

