% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{copyData}
\alias{copyData}
\title{Copy tidy data to clipboard}
\usage{
copyData(df, sep = "\\t")
}
\arguments{
\item{df}{a data.frame, atmoic vector, or matrix}

\item{sep}{a character vector of length 1 indicating how the pasted columns
are separated}
}
\description{
Copy tidy data to clipboard
}
\details{
This function copies df to the clipboard, for easy pasting into 
word processors and spreadsheets (e.g., MS Excel).
}
\examples{

library(magrittr)
mtcars \%>\% 
lm(qsec ~ cyl + wt + hp, data = .) \%>\%
tidy \%>\%
copyData


}
