% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_tidiers.R
\name{cox.zph_tidiers}
\alias{cox.zph_tidiers}
\alias{glance.cox.zph}
\alias{tidy.cox.zph}
\title{Tidiers for Tests of Proportional Hazards Assumption}
\usage{
\method{tidy}{cox.zph}(x, global = TRUE, ...)

\method{glance}{cox.zph}(x, ...)
}
\arguments{
\item{x}{a "survdiff" object}

\item{global}{logical, whether to include the global test in the last row}

\item{...}{other arguments passed to/from other methods, currently ignored}
}
\value{
All tidying methods return a data.frame without rownames, whose
structure depends on the method chosen.

\code{tidy} on "cox.zph" objects returns a data frame with the following columns:
\item{term}{model term}
\item{rho}{correlation between transformed survival time and the scaled Schoenfeld residuals}
\item{statistic}{Chi-square test}
\item{p.value}{two-sided p-value}
If \code{global} is \code{TRUE} (default) last row if contains the global 
test (\code{term} is equal to \code{"GLOBAL"} and \code{rho} is \code{NA}).
If \code{global} is \code{FALSE} then this row is dropped.

\code{glance} returns a data frame with the following columns:
\item{statistic}{Chi-square test}
\item{p.value}{p-value}.
}
\description{
Tidiers for Tests of Proportional Hazards Assumption
}
\examples{
if( require("survival") ) {
     # from `help("cox.zph")`
     fit <- coxph(Surv(futime, fustat) ~ age + ecog.ps,
         data=ovarian)
     zph <- cox.zph(fit)
     tidy(zph)
}
}
\seealso{
\code{\link[survival]{cox.zph}}
}

