% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletRegModel_tidiers.R
\name{DirichletReg_tidiers}
\alias{DirichletReg_tidiers}
\alias{tidy.DirichletRegModel}
\alias{augment.DirichletRegModel}
\alias{glance.DirichletRegModel}
\title{Tidying methods for a Dirichlet regression model}
\usage{
\method{tidy}{DirichletRegModel}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

\method{augment}{DirichletRegModel}(x, data = x$data, newdata = NULL,
  type.predict = c("mu", "alpha", "phi"), type.residuals = c("standardized",
  "composite", "raw"), ...)

\method{glance}{DirichletRegModel}(x, ...)
}
\arguments{
\item{x}{DirichletRegModel object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for alternative parametrization)}

\item{...}{extra arguments (not used)}

\item{data}{Original data, defaults to the extracting it from the model}

\item{newdata}{If provided, performs predictions on the new data}

\item{type.predict}{character vector. Values to predict; passed on to
\code{\link{predict.DirichletReg}}. Valid options are "mu", "alpha" and "phi"}

\item{type.residuals}{character vector. Type of residuals to compute; passed on to
\code{\link{residuals.DirichletReg}} Valid options are "standardized", "composite" and "raw"}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy.DirichletRegModel} returns one row for each coefficient, with six columns:
  \item{response}{Which response column the coefficients correspond to}
  \item{term}{The term in the linear model being estimated and tested}
  \item{estimate}{The estimated coefficient}
  \item{std.error}{The standard error from the linear model}
  \item{statistic}{t-statistic}
  \item{p.value}{two-sided p-value}


If \code{conf.int=TRUE}, it also includes columns for \code{conf.low} and
\code{conf.high}, computed with \code{\link{confint}}.

When \code{newdata} is not supplied \code{augment.DirichletRegModel} returns
one row for each observation, with seven columns added to the original
data:
  \item{.fitted}{Fitted values of model (mu for extected values, alpha values and phi for precision values )}
  \item{.resid}{Residuals of type "standardized", "composite", or "raw"}


When \code{newdata} is supplied, \code{augment.DirichletRegModel} returns one row for each
observation, with multiple columns added to the new data:
  \item{.fitted}{Fitted values of model}

\code{glance.lm} returns a one-row data.frame with the columns
  \item{logLik}{the data's log-likelihood under the model}
  \item{AIC}{the Akaike Information Criterion}
  \item{BIC}{the Bayesian Information Criterion}
  \item{deviance}{deviance}
  \item{df.residual}{residual degrees of freedom}
}
\description{
These methods tidy the coefficients of a Dirichlet regression models into a summary,
augment the original data with information on the fitted values and
residuals, and construct a one-row glance of the model's statistics.
}
\details{
If you have missing values in your model data, you may need to refit
the model with \code{na.action = na.exclude}.

If \code{conf.int=TRUE}, the confidence interval is computed with
the \code{\link{confint}} function.

When the modeling was performed with \code{na.action = "na.omit"}
(as is the typical default), rows with NA in the initial data are omitted
entirely from the augmented data frame. When the modeling was performed
with \code{na.action = "na.exclude"}, one should provide the original data
as a second argument, at which point the augmented data will contain those
rows (typically with NAs in place of the new columns). If the original data
is not provided to \code{augment} and \code{na.action = "na.exclude"}, a
warning is raised and the incomplete rows are dropped.
}
\examples{
library(DirichletReg)
ALake <- ArcticLake
ALake$Y <- DR_data(ALake[,1:3])
# fit a quadratic Dirichlet regression models ("common")
res1 <- DirichReg(Y ~ depth + I(depth^2), ALake)
}
\seealso{
\code{\link{summary.lm}}

\link{na.action}
}
