% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatr-tidiers.R
\name{estimatr_tidiers}
\alias{estimatr_tidiers}
\alias{tidy.lm_robust}
\alias{tidy.iv_robust}
\alias{tidy.difference_in_means}
\alias{tidy.horvitz_thompson}
\title{Tidy an estimatr object}
\usage{
\method{tidy}{lm_robust}(object, ...)

\method{tidy}{iv_robust}(object, ...)

\method{tidy}{difference_in_means}(object, digits = NULL)

\method{tidy}{horvitz_thompson}(object, digits = NULL)
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns for coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom, and the
name of the outcome variable
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[estimatr:lm_robust]{estimatr::lm_robust()}}, \code{\link[estimatr:iv_robust]{estimatr::iv_robust()}},  \code{\link[estimatr:difference_in_means]{estimatr::difference_in_means()}}, \code{\link[estimatr:horvitz_thompson]{estimatr::horvitz_thompson()}}
}
\concept{estimatr tidiers}
