% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factanal_tidiers.R
\name{factanal_tidiers}
\alias{factanal_tidiers}
\alias{tidy.factanal}
\alias{augment.factanal}
\alias{glance.factanal}
\title{Tidying methods for a factor analysis}
\usage{
\method{tidy}{factanal}(x, ...)

\method{augment}{factanal}(x, data, ...)

\method{glance}{factanal}(x, ...)
}
\arguments{
\item{x}{\code{\link{factanal}} object}

\item{...}{Additional arguments, not used}

\item{data}{Original data}
}
\value{
All tidying methods return a \code{\link[tibble]{tibble}} without
  rownames.

\code{tidy.factanal} returns one row for each variable used in the 
  analysis and the following columns:
  \item{variable}{The variable being estimated in the factor analysis}
  \item{uniqueness}{Proportion of residual, or unexplained variance}
  \item{flX}{Factor loading of term on factor X. There will be as many columns
  of this format as there were factors fitted.}

When \code{data} is not supplied \code{augment.factanal} returns one 
  row for each observation, with a factor score column added for each factor 
  X, (\code{.fsX}). This is because \code{\link{factanal}}, unlike other 
  stats methods like \code{\link{lm}}, does not retain the original data.

When \code{data} is supplied, \code{augment.factanal} returns one row for
each observation, with a factor score column added for each factor X,
(\code{.fsX}).

\code{glance.factanal} returns a one-row data.frame with the columns:
  \item{n.factors}{The number of fitted factors}
  \item{total.variance}{Total cumulative proportion of variance accounted for by all factors}
  \item{statistic}{Significance-test statistic}
  \item{p.value}{p-value from the significance test, describing whether the
  covariance matrix estimated from the factors is significantly different
  from the observed covariance matrix}
  \item{df}{Degrees of freedom used by the factor analysis}
  \item{n}{Sample size used in the analysis}
  \item{method}{The estimation method; always Maximum Likelihood, "mle"}
  \item{converged}{Whether the factor analysis converged}
}
\description{
These methods tidy the factor loadings of a factor analysis, conducted via 
\code{\link{factanal}}, into a summary, augment the original data with factor
scores, and construct a one-row glance of the model's statistics.
}
\examples{

mod <- factanal(mtcars, 3, scores = "regression")

glance(mod)
tidy(mod)
augment(mod)
augment(mod, mtcars)  # Must include original data if desired

}
