% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactoMineR_tidiers.R
\name{FactoMineR_tidiers}
\alias{FactoMineR_tidiers}
\alias{augment.MCA}
\alias{glance.MCA}
\alias{tidy.MCA}
\title{tidying methods for objects produced by \pkg{FactoMineR}}
\usage{
\method{tidy}{MCA}(x, var.sup = NULL, n = 5, ...)

\method{augment}{MCA}(x, data = x$call$X, ind.sup = NULL, n = 5, ...)

\method{glance}{MCA}(x, n = 5)
}
\arguments{
\item{x}{an object of class \code{MCA} or \code{PCA}}

\item{var.sup}{supplementary variables. A data frame with the same number of rows as the original dataframes, and discrete variables}

\item{n}{number of dimensions to take into account}

\item{...}{extra arguments (not used)}

\item{data}{original data this was fitted on}

\item{ind.sup}{a dataframe with supplementary individuals. It must have the same (active) variables as the original dataframe}

\item{n}{number of dimensions to take into account}

\item{n}{number of dimensions to retain}

\item{...}{extra arguments (not used)}
}
\value{
\code{augment} returns one row for each original observation, and for the supplemnetary individuals if there are some,
with columns (each prepended by a .) added. They are called .Dim followed by the numer of the dimension.

\code{glance} returns one row with each column corresponding to the eigenvalue of the dimension
}
\description{
These methods provide some summaries of objects produced by FactoMineR, such as MCA or PCA.
}
\details{
When the modeling was performed with \code{na.action = "na.omit"}
(as is the typical default), rows with NA in the initial data are omitted
entirely from the augmented data frame. When the modeling was performed
with \code{na.action = "na.exclude"}, one should provide the original data
as a second argument, at which point the augmented data will contain those
rows (typically with NAs in place of the new columns). If the original data
is not provided to \code{augment} and \code{na.action = "na.exclude"}, a
warning is raised and the incomplete rows are dropped.
}
\examples{

}
\seealso{
\link{na.action}
}

