% Generated by roxygen2 (4.0.2): do not edit by hand
\name{finish_glance}
\alias{finish_glance}
\title{Add logLik, AIC, BIC, and other common measurements to a glance of
a prediction}
\usage{
finish_glance(ret, x)
}
\arguments{
\item{ret}{a one-row data frame (a partially complete glance)}

\item{x}{the prediction model}
}
\value{
a data frame with additional columns added
}
\description{
A helper function for several functions in the glance generic. Methods
such as logLik, AIC, and BIC are defined for many prediction
objects, such as lm, glm, and nls. This is a helper function that adds
them to a glance data.frame can be performed. If any of them cannot be
computed, it fails quietly.
}

