% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-0-7-0.R
\name{fix_data_frame}
\alias{fix_data_frame}
\title{Ensure an object is a data frame, with rownames moved into a column}
\usage{
fix_data_frame(x, newnames = NULL, newcol = "term")
}
\arguments{
\item{x}{a data.frame or matrix}

\item{newnames}{new column names, not including the rownames}

\item{newcol}{the name of the new rownames column}
}
\value{
a data.frame, with rownames moved into a column and new column
names assigned
}
\description{
This function is deprecated as of broom 0.7.0 and will be removed from
a future release. Please see \code{tibble::as_tibble}.
}
\seealso{
Other deprecated: 
\code{\link{bootstrap}()},
\code{\link{confint_tidy}()},
\code{\link{data.frame_tidiers}},
\code{\link{finish_glance}()},
\code{\link{summary_tidiers}},
\code{\link{tidy.density}()},
\code{\link{tidy.dist}()},
\code{\link{tidy.ftable}()},
\code{\link{tidy.numeric}()}
}
\concept{deprecated}
