% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-aareg-tidiers.R
\name{glance.aareg}
\alias{glance.aareg}
\title{Glance at a(n) aareg object}
\usage{
\method{glance}{aareg}(x, ...)
}
\arguments{
\item{x}{An \code{aareg} object returned from \code{\link[survival:aareg]{survival::aareg()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{statistic}{chi-squared statistic}
\item{p.value}{p-value based on chi-squared statistic}
\item{df}{degrees of freedom used by coefficients}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[survival:aareg]{survival::aareg()}}

Other aareg tidiers: \code{\link{tidy.aareg}}

Other survival tidiers: \code{\link{augment.coxph}},
  \code{\link{augment.survreg}}, \code{\link{glance.cch}},
  \code{\link{glance.coxph}}, \code{\link{glance.pyears}},
  \code{\link{glance.survdiff}},
  \code{\link{glance.survexp}},
  \code{\link{glance.survfit}},
  \code{\link{glance.survreg}}, \code{\link{tidy.aareg}},
  \code{\link{tidy.cch}}, \code{\link{tidy.coxph}},
  \code{\link{tidy.pyears}}, \code{\link{tidy.survdiff}},
  \code{\link{tidy.survexp}}, \code{\link{tidy.survfit}},
  \code{\link{tidy.survreg}}
}
\concept{aareg tidiers}
\concept{survival tidiers}
