% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-arima-tidiers.R
\name{glance.Arima}
\alias{glance.Arima}
\title{Glance at a(n) Arima object}
\usage{
\method{glance}{Arima}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Arima} created by \code{\link[stats:arima]{stats::arima()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{sigma}{the square root of the estimated residual variance}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[stats:arima]{stats::arima()}}

Other Arima tidiers: \code{\link{tidy.Arima}}
}
\concept{Arima tidiers}
