% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg-tidiers.R
\name{glance.betareg}
\alias{glance.betareg}
\title{Glance at a(n) betareg object}
\usage{
\method{glance}{betareg}(x, ...)
}
\value{
\code{glance} returns a one-row tibble with columns:
\item{pseudo.r.squared}{the deviance of the null model}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{df.residual}{residual degrees of freedom}
\item{df.null}{degrees of freedom under the null}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[betareg:betareg]{betareg::betareg()}}
}
