% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm-tidiers.R
\name{glance.biglm}
\alias{glance.biglm}
\title{Glance at a(n) biglm object}
\usage{
\method{glance}{biglm}(x, ...)
}
\value{
\code{glance.biglm} returns a one-row data frame, with columns
\item{r.squared}{The percent of variance explained by the model}
\item{AIC}{the Akaike Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[biglm:biglm]{biglm::biglm()}}, \code{\link[biglm:bigglm]{biglm::bigglm()}}

Other biglm tidiers: \code{\link{tidy.biglm}}
}
\concept{biglm tidiers}
