% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-cch-tidiers.R
\name{glance.cch}
\alias{glance.cch}
\title{Glance at a(n) cch object}
\usage{
\method{glance}{cch}(x, ...)
}
\arguments{
\item{x}{An \code{cch} object returned from \code{\link[survival:cch]{survival::cch()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{score}{score}
\item{rscore}{rscore}
\item{p.value}{p-value from Wald test}
\item{iter}{number of iterations}
\item{n}{number of predictions}
\item{nevent}{number of events}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[survival:cch]{survival::cch()}}

Other cch tidiers: \code{\link{glance.survfit}},
  \code{\link{tidy.cch}}

Other survival tidiers: \code{\link{augment.coxph}},
  \code{\link{augment.survreg}},
  \code{\link{glance.aareg}}, \code{\link{glance.coxph}},
  \code{\link{glance.pyears}},
  \code{\link{glance.survdiff}},
  \code{\link{glance.survexp}},
  \code{\link{glance.survfit}},
  \code{\link{glance.survreg}}, \code{\link{tidy.aareg}},
  \code{\link{tidy.cch}}, \code{\link{tidy.coxph}},
  \code{\link{tidy.pyears}}, \code{\link{tidy.survdiff}},
  \code{\link{tidy.survexp}}, \code{\link{tidy.survfit}},
  \code{\link{tidy.survreg}}
}
\concept{cch tidiers}
\concept{survival tidiers}
