% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-cv-glmnet-tidiers.R
\name{glance.cv.glmnet}
\alias{glance.cv.glmnet}
\title{Glance at a(n) cv.glmnet object}
\usage{
\method{glance}{cv.glmnet}(x, ...)
}
\arguments{
\item{x}{A \code{cv.glmnet} object returned from \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with one-row with columns:
\item{lambda.min}{The value of the penalization parameter lambda that
achieved minimum loss as estimated by cross validation.}
\item{lambda.1se}{The value of the penalization parameter lambda that
results in the sparsest model while remaining within one standard
error of the minimum loss.}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}

Other glmnet tidiers: \code{\link{glance.glmnet}},
  \code{\link{tidy.cv.glmnet}}, \code{\link{tidy.glmnet}}
}
