% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-fitdistr-tidiers.R
\name{glance.fitdistr}
\alias{glance.fitdistr}
\title{Glance at a(n) fitdistr object}
\usage{
\method{glance}{fitdistr}(x, ...)
}
\arguments{
\item{x}{A \code{fitdistr} object returned by \code{\link[MASS:fitdistr]{MASS::fitdistr()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{n}{Number of observations used in estimation}
\item{logLik}{log-likelihood of estimated data}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[MASS:fitdistr]{MASS::fitdistr()}}

Other fitdistr tidiers: \code{\link{tidy.fitdistr}}
}
