% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest-tidiers.R
\name{glance.fixest}
\alias{glance.fixest}
\title{Glance at a(n) fixest object}
\usage{
\method{glance}{fixest}(x, ...)
}
\arguments{
\item{x}{A \code{fixest} object returned from any of the \code{fixest} estimators}

\item{...}{Additional arguments passed to \code{summary} and \code{confint}. Important
arguments are \code{se} and \code{cluster}. See \code{\link[fixest:summary.fixest]{summary.fixest}}.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
\itemize{
\item logLik The log-likelihood of the model. \code{\link[stats:logLik]{stats::logLik()}} may be a useful
reference.
\item AIC Akaike's Information Criterion for the model.
\item BIC Bayesian Information Criterion for the model.
\item nobs Number of observations used.
\item sigma (\code{feols} models only)
\item r.squared (\code{feols} models only)
\item adj.r.squared Adjusted R squared statistic, which is like the R squared
statistic except taking degrees of freedom into account. (\code{feols} models only)
\item within.r.squared R squared within fixed-effect groups (\code{feols} models only)
\item pseudo.r.squared Like the R squared statistic, but for situations when the
R squared statistic isn't defined. (non-\code{feols} models only)
}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.
}
\note{
The columns of the result depend on the type of model estimated.
}
\examples{
\donttest{
library(fixest)
gravity <- feols(log(Euros) ~ log(dist_km) | Origin + Destination + Product + Year, trade)

tidy(gravity)
glance(gravity)
augment(gravity, trade)

## To get robust or clustered SEs, users can either:
# 1) Feed tidy() a summary.fixest object that has already accepted these arguments
gravity_summ <- summary(gravity, cluster = c("Product", "Year"))
tidy(gravity_summ, conf.int = TRUE)
# 2) Or, specify the arguments directly in the tidy() call
tidy(gravity, conf.int = TRUE, cluster = c("Product", "Year"))
tidy(gravity, conf.int = TRUE, se = "threeway")
# etc.

## The other fixest methods all working similarly. For example:
gravity_pois <- feglm(Euros ~ log(dist_km) | Origin + Destination + Product + Year, trade)
tidy(gravity_pois)
glance(gravity_pois)
augment(gravity_pois, trade)
}

}
