% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-glm-tidiers.R
\name{glance.glm}
\alias{glance.glm}
\title{Glance at a(n) glm object}
\usage{
\method{glance}{glm}(x, ...)
}
\arguments{
\item{x}{A \code{glm} object returned from \code{\link[stats:glm]{stats::glm()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{null.deviance}{the deviance of the null model}
\item{df.null}{the residual degrees of freedom for the null model}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\examples{

g <- glm(am ~ mpg, mtcars, family = "binomial")
glance(g)

}
\seealso{
\code{\link[stats:glm]{stats::glm()}}

Other lm tidiers: \code{\link{augment.glm}},
  \code{\link{tidy.glm}}
}
\concept{lm tidiers}
