% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-glmnet-tidiers.R
\name{glance.glmnet}
\alias{glance.glmnet}
\title{Glance at a(n) glmnet object}
\usage{
\method{glance}{glmnet}(x, ...)
}
\arguments{
\item{x}{A \code{glmnet} object returned from \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{nulldev}{null deviance}
\item{npasses}{total passes over the data across all lambda values}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[glmnet:glmnet]{glmnet::glmnet()}}

Other glmnet tidiers: \code{\link{glance.cv.glmnet}},
  \code{\link{tidy.cv.glmnet}}, \code{\link{tidy.glmnet}}
}
