% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-glmrob-tidiers.R
\name{glance.glmRob}
\alias{glance.glmRob}
\title{Glance at a(n) glmRob object}
\usage{
\method{glance}{glmRob}(x, ...)
}
\arguments{
\item{x}{A \code{glmRob} object returned from \code{\link[robust:glmRob]{robust::glmRob()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{deviance}{Robust deviance}
\item{null.deviance}{Deviance under the null model}
\item{df.residual}{Number of residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[robust:glmRob]{robust::glmRob()}}

Other robust tidiers: \code{\link{augment.glmRob}},
  \code{\link{augment.lmRob}}, \code{\link{glance.lmRob}},
  \code{\link{tidy.glmRob}}, \code{\link{tidy.lmRob}}
}
\concept{robust tidiers}
