% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glance.htest}
\alias{glance.htest}
\title{one-row glance of an htset object}
\usage{
\method{glance}{htest}(x, ...)
}
\arguments{
\item{x}{htest object}

\item{...}{extra arguments, not used}
}
\description{
Returns the same output as \code{\link{tidy.htest}} (since it is already one-line long)
}
\seealso{
tidy.htest

\code{\link{tidy.htest}}
}

