% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glance.kmeans}
\alias{glance.kmeans}
\title{one-row glance of a kmeans object}
\usage{
\method{glance}{kmeans}(x, ...)
}
\arguments{
\item{x}{kmeans object}

\item{...}{extra arguments, not used}
}
\value{
a one-row data.frame with the columns
\itemize{
\item{totss}{The total sum of squares}
\item{tot.withinss}{The total within-cluster sum of squares}
\item{betweenss}{The total between-cluster sum of squares}
\item{iter}{The numbr of (outer) iterations}
}
}
\description{
Construct a one-row data frame summarizing a k-means object.
}
\seealso{
\code{\link{kmeans}}
}

