% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-lmrob-tidiers.R
\name{glance.lmRob}
\alias{glance.lmRob}
\title{Glance at a(n) lmRob object}
\usage{
\method{glance}{lmRob}(x, ...)
}
\arguments{
\item{x}{A \code{lmRob} object returned from \code{\link[robust:lmRob]{robust::lmRob()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{r.squared}{R-squared}
\item{deviance}{Robust deviance}
\item{sigma}{Residual scale estimate}
\item{df.residual}{Number of residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[robust:lmRob]{robust::lmRob()}}

Other robust tidiers: \code{\link{augment.glmRob}},
  \code{\link{augment.lmRob}}, \code{\link{glance.glmRob}},
  \code{\link{tidy.glmRob}}, \code{\link{tidy.lmRob}}
}
\concept{robust tidiers}
