% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistf-tidiers.R
\name{glance.logistf}
\alias{glance.logistf}
\title{Glance at a(n) logistf object}
\usage{
\method{glance}{logistf}(x, ...)
}
\arguments{
\item{x}{A \code{logistf} object returned from \code{\link[logistf:logistf]{logistf::logistf()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.level = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modeling
function. This includes the name of the modeling function or any
arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.

Glance returns the same number of columns regardless of whether the
model matrix is rank-deficient or not. If so, entries in columns
that no longer have a well-defined value are filled in with an \code{NA}
of the appropriate type.
}
\examples{
library(logistf)

dat <- data.frame(time = 1:12, 
                  event = c(0,1,0,1,1,1,1,1,1,1,1,1))
mod <- logistf::logistf(event ~time, dat)
tidy(mod)
glance(mod)

}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{df}{Degrees of freedom used by the model.}
  \item{nobs}{Number of observations used.}

}
