% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust-tidiers.R
\name{glance.Mclust}
\alias{glance.Mclust}
\title{Glance at a(n) Mclust object}
\usage{
\method{glance}{Mclust}(x, ...)
}
\arguments{
\item{x}{An \code{Mclust} object return from \code{\link[mclust:Mclust]{mclust::Mclust()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{model}{A character string denoting the model at which the optimal
BIC occurs}
\item{n}{The number of observations in the data}
\item{G}{The optimal number of mixture components}
\item{BIC}{The optimal BIC value}
\item{logLik}{The log-likelihood corresponding to the optimal BIC}
\item{df}{The number of estimated parameters}
\item{hypvol}{If the other model contains a noise component, the
value of the hypervolume parameter. Otherwise \code{NA}.}
}
\description{
Description of what a \code{glance} method does goes here.
}
