% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls-mvr-tidiers.R
\name{glance.mvr}
\alias{glance.mvr}
\title{Glance at a(n) mvr object}
\usage{
\method{glance}{mvr}(x, ncomp = x$ncomp, ...)
}
\arguments{
\item{x}{An \code{mvr} object such as those created by \code{\link[pls:mvr]{pls::mvr()}}, \code{\link[pls:plsr]{pls::plsr()}}, \code{\link[pls:cppls]{pls::cppls()}} and \code{\link[pls:pcr]{pls::pcr()}}.}

\item{ncomp}{The number of components to include in the model. Ignored if comps is specified.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.
}
\examples{
library(pls)
data(yarn) # Single-response model
yarn.pls <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
glance(yarn.pls)
tidy(yarn.pls)
augment(yarn.pls)

data(oliveoil) # Multi-response model
sens.pls <- plsr(sensory ~ chemical, ncomp = 4, scale = TRUE, data = oliveoil)
glance(sens.pls)
tidy(sens.pls)
augment(sens.pls)

set.seed(42)
test_id <- sample(1:nrow(yarn), nrow(yarn) / 3)
yarn.pls.train <- plsr(density ~ NIR, 3, data = slice(yarn, -test_id))
augment(yarn.pls.train, newdata = slice(yarn, test_id)) \%>\%
  mutate(.rownames = test_id) \%>\%
  ggplot(aes(.rownames, .t.squared[, 3])) +
  geom_point() +
  ggtitle("How far each sample is from the center of the model.")
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[pls:R2]{pls::R2()}}

Other mvr tidiers: \code{\link{augment.mvr}},
  \code{\link{tidy.mvr}}
}
\concept{mvr tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{r.squared}{R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.}

}
