% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glance.nls}
\alias{glance.nls}
\title{one-row glance of an nls object}
\usage{
\method{glance}{nls}(x, ...)
}
\arguments{
\item{x}{nls object}

\item{...}{extra arguments, not used}
}
\value{
a one-row data.frame with the columns
\itemize{
\item{sigma}{The square root of the estimated residual variance}
\item{isConv}{Whether the fit successfully converged}
\item{finTol}{The achieved convergence tolerance}
}
}
\description{
Construct a one-row data frame summarizing an nls object, including
the standard deviation of the residuals, and whether it has converged.
}
\seealso{
\code{\link{nls}} and \code{\link{summary.nls}}
}

