% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcutt-tidiers.R
\name{glance.orcutt}
\alias{glance.orcutt}
\title{Glance at a(n) orcutt object}
\usage{
\method{glance}{orcutt}(x, ...)
}
\arguments{
\item{x}{An \code{orcutt} object returned from \code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{r.squared}{R-squared}
\item{adj.r.squared}{Adjusted R-squared}
\item{rho}{Spearman's rho autocorrelation}
\item{number.interaction}{Number of interactions}
\item{dw.original}{Durbin-Watson statistic of original fit}
\item{p.value.original}{P-value of original Durbin-Watson statistic}
\item{dw.transformed}{Durbin-Watson statistic of transformed fit}
\item{p.value.transformed}{P-value of autocorrelation after transformation}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}

Other orcutt tidiers: \code{\link{tidy.orcutt}}
}
