% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polca-tidiers.R
\name{glance.poLCA}
\alias{glance.poLCA}
\title{Augment data with information from a(n) poLCA object}
\usage{
\method{glance}{poLCA}(x, ...)
}
\arguments{
\item{x}{A \code{poLCA} object returned from \code{\link[poLCA:poLCA]{poLCA::poLCA()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{g.squared}{The likelihood ratio/deviance statistic}
\item{chi.squared}{The Pearson Chi-Square goodness of fit statistic
for multiway tables}
\item{df}{Number of parameters estimated, and therefore degrees of
freedom used}
\item{df.residual}{Number of residual degrees of freedom left}
}
\description{
Description of what a \code{augment} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[poLCA:poLCA]{poLCA::poLCA()}}

Other poLCA tidiers: \code{\link{augment.poLCA}},
  \code{\link{tidy.poLCA}}
}
\concept{poLCA tidiers}
