% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glance.R
\name{glance}
\alias{glance}
\title{Construct a single row summary "glance" of a model, fit, or other
object}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\description{
glance methods always return either a one-row data frame, or NULL
}

