% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-ridgelm-tidiers.R
\name{glance.ridgelm}
\alias{glance.ridgelm}
\title{Glance at a(n) ridgelm object}
\usage{
\method{glance}{ridgelm}(x, ...)
}
\arguments{
\item{x}{A \code{ridgelm} object returned from \code{\link[MASS:lm.ridge]{MASS::lm.ridge()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{kHKB}{modified HKB estimate of the ridge constant}
\item{kLW}{modified L-W estimate of the ridge constant}
\item{lambdaGCV}{choice of lambda that minimizes GCV}
}
\description{
Description of what a \code{glance} method does goes here.
}
\details{
This is similar to the output of \code{select.ridgelm}, but it is
returned rather than printed.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[MASS:select.ridgelm]{MASS::select.ridgelm()}}, \code{\link[MASS:lm.ridge]{MASS::lm.ridge()}}

Other ridgelm tidiers: \code{\link{tidy.ridgelm}}
}
\concept{ridgelm tidiers}
