% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{glance.rma}
\alias{glance.rma}
\title{Construct a one-row summary of meta-analysis model fit statistics.}
\usage{
\method{glance}{rma}(x, ...)
}
\arguments{
\item{x}{An \code{rma} created by the \code{metafor} package.}

\item{...}{Additional arguments}
}
\value{
a \code{tibble}
}
\description{
\code{glance()} computes a one-row summary of  meta-analysis objects,
including estimates of heterogenity and model fit.
}
\examples{

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

glance(meta_analysis)

}
