% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg-rq-tidiers.R
\name{glance.rq}
\alias{glance.rq}
\title{Glance at a(n) rq object}
\usage{
\method{glance}{rq}(x, ...)
}
\arguments{
\item{x}{An \code{rq} object returned from \code{\link[quantreg:rq]{quantreg::rq()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{tau}{quantile estimated}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\details{
Only models with a single \code{tau} value may be passed.
For multiple values, please use a \code{\link[purrr:map]{purrr::map()}} workflow instead, e.g.\preformatted{taus \%>\%
  map(function(tau_val) rq(y ~ x, tau = tau_val)) \%>\%
  map_dfr(glance)
}
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[quantreg:rq]{quantreg::rq()}}

Other quantreg tidiers: \code{\link{augment.nlrq}},
  \code{\link{augment.rqs}}, \code{\link{augment.rq}},
  \code{\link{glance.nlrq}}, \code{\link{tidy.nlrq}},
  \code{\link{tidy.rqs}}, \code{\link{tidy.rq}}
}
\concept{quantreg tidiers}
