% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-tidiers.R
\name{glance.svyglm}
\alias{glance.svyglm}
\title{Glance at a(n) svyglm object}
\usage{
\method{glance}{svyglm}(x, maximal = x, ...)
}
\arguments{
\item{x}{A \code{svyglm} object returned from \code{\link[survey:svyglm]{survey::svyglm()}}.}

\item{maximal}{A \code{svyglm} object corresponding to the maximal
model against which to compute the BIC. See Lumley and Scott
(2015) for details. Defaults to \code{x}, which is equivalent
to not using a maximal model.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.
}
\examples{

library(survey)

set.seed(123)
data(api)

# survey design
dstrat <-
  svydesign(
    id =  ~ 1,
    strata =  ~ stype,
    weights =  ~ pw,
    data = apistrat,
    fpc =  ~ fpc
  )

# model
m <- survey::svyglm(
  formula = sch.wide ~ ell + meals + mobility,
  design = dstrat,
  family = quasibinomial()
)

glance(m)

}
\references{
Lumley T, Scott A (2015). AIC and BIC for modelling with complex
survey data. \emph{Journal of Survey Statistics and Methodology}, 3(1).
\url{https://doi.org/10.1093/jssam/smu021}.
}
\seealso{
\code{\link[survey:svyglm]{survey::svyglm()}}, \code{\link[stats:glm]{stats::glm()}}, \link[survey:anova.svyglm]{survey::anova.svyglm}

Other lm tidiers: 
\code{\link{augment.glm}()},
\code{\link{augment.lm}()},
\code{\link{glance.glm}()},
\code{\link{glance.lm}()},
\code{\link{tidy.glm}()},
\code{\link{tidy.lm.beta}()},
\code{\link{tidy.lm}()},
\code{\link{tidy.mlm}()}
}
\concept{lm tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{AIC}{Akaike's Information Criterion for the model.}
  \item{BIC}{Bayesian Information Criterion for the model.}
  \item{deviance}{Deviance of the model.}
  \item{df.null}{Degrees of freedom used by the null model.}
  \item{df.residual}{Residual degrees of freedom.}
  \item{null.deviance}{Deviance of the null model.}

}
