% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-tidiers.R
\name{glance.xgb.Booster}
\alias{glance.xgb.Booster}
\title{Tidy a(n) xgb.Booster object}
\usage{
\method{glance}{xgb.Booster}(x)
}
\arguments{
\item{x}{an xgb.Booster object}
}
\value{
A summary of evaluation metric results for at the last iteration of the xgb.Booster object, and, if \code{early_stopping_rounds} was specified during training, the results at the best iteration.
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\examples{

mdl <- copy(iris)
features <- setdiff(names(iris), "Species")
mdl[['Species']] <- as.integer(mdl[['Species']]) - 1
dtrain <- xgb.DMatrix(as.matrix(mdl[1:120, features]), label = mdl[['Species']][1:120])
dvalidate <- xgb.DMatrix(as.matrix(mdl[121:150, features]), label = mdl[['Species']][121:150])
xgb_params <- list(
  objective = 'multi:softmax', 
  eval_metric = c('mlogloss', 'merror'), 
  num_class = 3
)
fit <- xgb.train(
  data = dtrain, 
  params = xgb_params, 
  watchlist = list(train = dtrain, validate = dvalidate), 
  nrounds = 50, 
  early_stopping_rounds = 5, 
  verbose = 0
)
glance(fit)

}
\seealso{
\code{\link[xgboost:xgb.train]{xgboost::xgb.train()}}, \code{\link[=xgboost:xgboost]{xgboost:xgboost()}}
}
