% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-optim-tidiers.R
\name{glance_optim}
\alias{glance_optim}
\alias{glance.optim}
\title{Tidying methods for singular value decomposition lists}
\usage{
glance_optim(x, ...)
}
\arguments{
\item{x}{A list returned from \code{\link[stats:optim]{stats::optim()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{value}{minimized or maximized output value}
\item{function.count}{number of calls to \code{fn}}
\item{gradient.count}{number of calls to \code{gr}}
\item{convergence}{convergence code representing the error state}
}
\description{
These methods tidy the U, D, and V matrices returned by the
\code{\link[=svd]{svd()}} functions into a tidy format. Because \code{svd} returns
lists without classes, this function has to be called by
\code{\link[=tidy.list]{tidy.list()}} when it recognizes a list as an SVD.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[=optim]{optim()}}

Other list tidiers: \code{\link{list_tidiers}},
  \code{\link{tidy_irlba}}, \code{\link{tidy_optim}},
  \code{\link{tidy_svd}}, \code{\link{tidy_xyz}}
}
\concept{list tidiers}
