% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_tidiers.R
\docType{class}
\name{h2o_tidiers}
\alias{h2o_tidiers}
\alias{tidy,H2OBinomialModel-method}
\alias{tidy,H2ORegressionModel-method}
\alias{augment,H2OBinomialModel-method}
\alias{augment,H2ORegressionModel-method}
\alias{glance,H2OModel-method}
\title{Tidying methods for H2O models}
\usage{
\S4method{tidy}{H2OBinomialModel}(x, exponentiate = FALSE)

\S4method{tidy}{H2ORegressionModel}(x, exponentiate = FALSE)

\S4method{augment}{H2OBinomialModel}(x,
  data = h2o::h2o.getFrame(x@allparameters$training_frame), newdata = NULL,
  collect = TRUE, ...)

\S4method{augment}{H2ORegressionModel}(x,
  data = h2o::h2o.getFrame(x@allparameters$training_frame), newdata = NULL,
  collect = TRUE, ...)

\S4method{glance}{H2OModel}(x, on = "train", ...)
}
\arguments{
\item{x}{An H2OModel object}

\item{exponentiate}{For GLM, whether to exponentiate the coefficient estimates (typical for logistic regression)}

\item{data}{H2OFrame on which the model is trained on}

\item{newdata}{If provided, performs predictions on the new data; must be an H2OFrame object}

\item{collect}{Whether to collect the output to a local R data.frame; defaults to \code{TRUE}}

\item{on}{The dataset on which model metrics will be evaluated; one of \code{"train"}, \code{"valid"}, and \code{"xval"}; defaults to \code{"train"}}
}
\value{
\code{tidy} for H2O GLM models returns one row for each coefficient. Note that p-values and associated statistics are only included if \code{compute_p_values} was set to \code{TRUE} during model fitting
  \item{term}{The term in the linear model being estimated and tested}
  \item{estimate}{The estimated coefficient}
  \item{std.error}{The standard error from the GLM}
  \item{statistic}{Z-score}
  \item{p.value}{p-value}

\code{augment} for H2O GLM models returns a data frame (or H2OFrame if \code{collect = FALSE}) with the following columns appended to \code{data} or \code{newdata}
  \item{.fitted}{Fitted values of model}
  \item{.se.fit}{Standard errors of fitted values}

\code{glance} for H2O models returns a one-row data.frame with columns representing various model metrics. In the case of GLM the following are outputted
  \item{r.squared}{The percent of variance explained by the model}
  \item{df.null}{Null degrees of freedom}
  \item{df.residual}{Residual degrees of freedom}
  \item{logLik}{the data's log-likelihood under the model}
  \item{AIC}{the Akaike Information Criterion}
  \item{deviance}{deviance}
  \item{MSE}{Mean squared error}
  \item{RMSE}{Root mean squared error}
  \item{AUC}{Area under the ROC curve}
  \item{logloss}{Log loss}
}
\description{
These methods summarize the results of H2O models into tidy forms.
}
\details{
By default, \code{augment} returns local data frame. For very large datasets, it may be desirable to set \code{collect = FALSE}, in which case the function returns a pointer to the appended H2OFrame.
}
