% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inflate}
\alias{inflate}
\title{Expand a dataset to include all factorial combinations of one or more
variables}
\usage{
inflate(.data, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{.data}{a tbl}

\item{...}{arguments}

\item{stringsAsFactors}{logical specifying if character vectors are
converted to factors.}
}
\value{
A tbl, grouped by the arguments in \code{...}
}
\description{
Expand a dataset to include all factorial combinations of one or more
variables
}

