% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_NAs}
\alias{insert_NAs}
\title{insert a row of NAs into a data frame wherever another data frame has NAs}
\usage{
insert_NAs(x, original)
}
\arguments{
\item{x}{data frame that has one row for each non-NA row in original}

\item{original}{data frame with NAs}
}
\description{
insert a row of NAs into a data frame wherever another data frame has NAs
}

