% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg_tidiers.R
\name{ivreg_tidiers}
\alias{ivreg_tidiers}
\alias{tidy.ivreg}
\alias{glance.ivreg}
\title{Tidiers for ivreg models}
\usage{
\method{tidy}{ivreg}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

\method{glance}{ivreg}(x, ...)
}
\arguments{
\item{x}{an "ivreg" object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals}

\item{...}{extra arguments, not used}

\item{data}{original dataset}
}
\value{
All tidying methods return a data.frame without rownames, whose
structure depends on the method chosen.

\code{tidy.plm} returns a data frame with one row per
coefficient, of the same form as \code{\link{tidy.lm}}.

\code{glance} returns a one-row data frame with columns
  \item{r.squared}{The percent of variance explained by the model}
  \item{adj.r.squared}{r.squared adjusted based on the degrees of freedom}
  \item{statistic}{Wald test statistic}
  \item{p.value}{p-value from the Wald test}
  \item{df}{Degrees of freedom used by the coefficients}
  \item{sigma}{The square root of the estimated residual variance}
  \item{df.residual}{residual degrees of freedom}
}
\description{
Tidiers for ivreg models
}
\examples{

if (require("AER", quietly = TRUE)) {
    data("CigarettesSW", package = "AER")
    CigarettesSW$rprice <- with(CigarettesSW, price/cpi)
    CigarettesSW$rincome <- with(CigarettesSW, income/population/cpi)
    CigarettesSW$tdiff <- with(CigarettesSW, (taxs - tax)/cpi)
    ivr <- ivreg(log(packs) ~ log(rprice) + log(rincome) | log(rincome) + tdiff + I(tax/cpi),
          data = CigarettesSW, subset = year == "1995")
    
    summary(ivr)
    
    tidy(ivr)
    tidy(ivr, conf.int = TRUE)
    tidy(ivr, conf.int = TRUE, conf.level = .9)
    
    head(augment(ivr))
    
    glance(ivr)
}

}
\seealso{
\code{\link{lm_tidiers}}
}
