% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_tidiers.R
\name{lavaan_tidiers}
\alias{lavaan_tidiers}
\alias{tidy.lavaan}
\alias{glance.lavaan}
\title{Tidying methods for lavaan models}
\usage{
\method{tidy}{lavaan}(x, conf.level = 0.95, ...)

\method{glance}{lavaan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan}, such as those from \code{\link[lavaan:cfa]{lavaan::cfa()}},
or \code{\link[lavaan:sem]{lavaan::sem()}}}

\item{conf.level}{Confidence level to use. Default is 0.95.}

\item{...}{For \code{tidy}, additional arguments passed to
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. Ignored for \code{glance}.`}
}
\value{
\code{glance} returns statistics measuring model fit. TODO: Document.
}
\description{
These methods tidy the coefficients of lavaan CFA and SEM models.
}
