% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lm-tidiers}
\alias{augment.lm}
\alias{glance.lm}
\alias{lm-tidiers}
\alias{tidy.lm}
\title{Tidying methods for a linear model}
\usage{
\method{tidy}{lm}(x, ...)

\method{augment}{lm}(x, data = x$model, ...)

\method{glance}{lm}(x, ...)
}
\arguments{
\item{x}{lm object}

\item{data}{Original data, defaults to the extracting it from the model}

\item{...}{extra arguments, not used}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy.lm} returns one row for each coefficient, with five columns:
  \item{term}{The term in the linear model being estimated and tested}
  \item{estimate}{The estimated coefficient}
  \item{stderror}{The standard error from the linear model}
  \item{statistic}{t-statistic}
  \item{p.value}{two-sided p-value}

\code{augment.lm} returns one row for each observation,
with six columns added to the original data:
  \item{.hat}{Diagonal of the hat matrix}
  \item{.sigma}{Estimate of residual standard deviation when
    corresponding observation is dropped from model}
  \item{.cooksd}{Cooks distance, \code{\link{cooks.distance}}}
  \item{.fitted}{Fitted values of model}
  \item{.resid}{Residuals}
  \item{.stdresid}{Standardised residuals}

\code{glance.lm} returns a one-row data.frame with the columns
  \item{r.squared}{The percent of variance explained by the model}
  \item{adj.r.squared}{r.squared adjusted based on the degrees of freedom}
  \item{sigma}{The square root of the estimated residual variance}
  \item{statistic}{F-statistic}
  \item{p.value}{p-value from the F test, describing whether the full regression is significant}
}
\description{
These methods tidy the coefficients of a linear model into a summary,
augment the original data with information on the fitted values and
residuals, and construct a one-row glance of the model's statistics.
}
\details{
If you have missing values in your model data, you may need to refit
the model with \code{na.action = na.exclude}.

Code and documentation for \code{augment.lm} originated in the
ggplot2 package, where it was called \code{fortify.lm}
}
\examples{
library(ggplot2)

mod <- lm(mpg ~ wt, data = mtcars)
head(augment(mod))
head(augment(mod, mtcars))

plot(mod, which = 1)
qplot(.fitted, .resid, data = mod) +
  geom_hline(yintercept = 0) +
  geom_smooth(se = FALSE)
qplot(.fitted, .stdresid, data = mod) +
  geom_hline(yintercept = 0) +
  geom_smooth(se = FALSE)
qplot(.fitted, .stdresid, data = augment(mod, mtcars),
  colour = factor(cyl))
qplot(mpg, .stdresid, data = augment(mod, mtcars), colour = factor(cyl))

plot(mod, which = 2)
# qplot(sample =.stdresid, data = mod, stat = "qq") + geom_abline()

plot(mod, which = 3)
qplot(.fitted, sqrt(abs(.stdresid)), data = mod) + geom_smooth(se = FALSE)

plot(mod, which = 4)
qplot(seq_along(.cooksd), .cooksd, data = mod, geom = "bar",
 stat="identity")

plot(mod, which = 5)
qplot(.hat, .stdresid, data = mod) + geom_smooth(se = FALSE)
ggplot(mod, aes(.hat, .stdresid)) +
  geom_vline(size = 2, colour = "white", xintercept = 0) +
  geom_hline(size = 2, colour = "white", yintercept = 0) +
  geom_point() + geom_smooth(se = FALSE)

qplot(.hat, .stdresid, data = mod, size = .cooksd) +
  geom_smooth(se = FALSE, size = 0.5)

plot(mod, which = 6)
ggplot(mod, aes(.hat, .cooksd)) +
  geom_vline(xintercept = 0, colour = NA) +
  geom_abline(slope = seq(0, 3, by = 0.5), colour = "white") +
  geom_smooth(se = FALSE) +
  geom_point()
qplot(.hat, .cooksd, size = .cooksd / .hat, data = mod) + scale_size_area()
}
\seealso{
\code{\link{summary.lm}}
}

