% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafor_tidiers.R
\name{metafor_tidiers}
\alias{glance.rma}
\alias{metafor_tidiers}
\alias{tidy.rma}
\title{Tidying methods for meta-analysis models (via the metafor package)}
\usage{
\method{tidy}{rma}(x, ...)

\method{glance}{rma}(x, ...)
}
\arguments{
\item{x}{model object returned by \code{\link[metafor]{rma}}}

\item{...}{ignored.}
}
\value{
All tidying methods return a data.frame without rownames, whose
structure depends on the method chosen.

\code{tidy.rma} returns a data frame with one row for each coefficient.
The columns include:

* b	- estimated coefficients of the model.

* se - standard errors of the coefficients.

* zval - test statistics of the coefficients.

* pval - p-values for the test statistics.

* ci.lb	- lower bound of the confidence intervals for the coefficients.

* ci.ub	- upper bound of the confidence intervals for the coefficients.

\code{glance.rma} returns a data frame with one row with common statistics.

* The columns include:

* tau2 - estimated amount of (residual) heterogeneity. Always 0 when method="FE".

* se.tau2 - estimated standard error of the estimated amount of (residual) heterogeneity.

* k - number of outcomes included in the model fitting.

* p - number of coefficients in the model (including the intercept).

* m - number of coefficients included in the omnibus test of coefficients.

* QE - test statistic for the test of (residual) heterogeneity.

* QEp - p-value for the test of (residual) heterogeneity.

* QM - test statistic for the omnibus test of coefficients.

* QMp - p-value for the omnibus test of coefficients.

* I2 - value of I^2. See print.rma.uni for more details.

* H2 - value of H^2. See print.rma.uni for more details.

* R2 - value of R^2. See print.rma.uni for more details.

* int.only - logical that indicates whether the model is an intercept-only model.
}
\description{
These methods include a tidy and a glance for a meta-analytic fixed- and 
random/mixed-effects models with or without moderators via linear (mixed-effects) 
models. See the documentation of the metafor-package for more details on these models.
}
\details{
A tidy method for the coefficients of a meta analysis.

A glance method for the meta analysis model's statistics.
}

