% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfx-tidiers.R
\name{mfx_tidiers}
\alias{mfx_tidiers}
\alias{tidy.mfx}
\alias{tidy.betamfx}
\alias{tidy.logitmfx}
\alias{tidy.negbinmfx}
\alias{tidy.poissonmfx}
\alias{tidy.probitmfx}
\alias{augment.betamfx}
\alias{augment.logitmfx}
\alias{augment.negbinmfx}
\alias{augment.poissonmfx}
\alias{augment.probitmfx}
\alias{glance.betamfx}
\alias{glance.logitmfx}
\alias{glance.negbinmfx}
\alias{glance.poissonmfx}
\alias{glance.probitmfx}
\title{Tidy a(n) mfx object}
\usage{
\method{tidy}{mfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{tidy}{betamfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{tidy}{logitmfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{tidy}{negbinmfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{tidy}{poissonmfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{tidy}{probitmfx}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{augment}{betamfx}(x, data = model.frame(x$fit), newdata = NULL,
  type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)

\method{augment}{logitmfx}(x, data = model.frame(x$fit),
  newdata = NULL, type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)

\method{augment}{negbinmfx}(x, data = model.frame(x$fit),
  newdata = NULL, type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)

\method{augment}{poissonmfx}(x, data = model.frame(x$fit),
  newdata = NULL, type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)

\method{augment}{probitmfx}(x, data = model.frame(x$fit),
  newdata = NULL, type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"), se_fit = FALSE, ...)

\method{glance}{betamfx}(x, ...)

\method{glance}{logitmfx}(x, ...)

\method{glance}{negbinmfx}(x, ...)

\method{glance}{poissonmfx}(x, ...)

\method{glance}{probitmfx}(x, ...)
}
\arguments{
\item{x}{A \code{betamfx}, \code{logitmfx}, \code{negbinmfx}, \code{poissonmfx}, or \code{probitmfx}  object.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.

The particular functions below provide generic tidy methods for
objects returned by the \code{mfx} package, preserving the calculated marginal
effects instead of the naive model coefficients. The returned tidy tibble
will also include an additional "atmean"column indicating how the marginal
effects were originally calculated (see Details below).
}
\details{
The \code{mfx} package provides methods for calculating marginal effects
for various generalised linear models (GLMs). Unlike standard linear models
(e.g. OLS regression), estimated model coefficients in a GLM cannot be
directly interpreted as marginal effects (i.e., the change in the response
variable predicted after a one unit change in one of the regressors). This
is because the estimated coefficients are multiplicative, dependent on both
the link function that was used for the estimation and any other variables
that were included in the model. When calculating marginal effects, users
must typically choose whether they want to use i) the average observation
in the data, or ii) the average of the sample marginal effects. See
\code{vignette("mfxarticle")} from the \code{mfx} vignette for more details.
}
\examples{
library(mfx)

## Using logit regression
mfx_logitreg <- logitmfx(am ~ cyl + hp + wt, atmean = T, data = mtcars)
tidy(mfx_logitreg)
tidy(mfx_logitreg, conf.int=T)
mfx_logitreg2 <- logitmfx(am ~ cyl + hp + wt, atmean = F, robust = T, data = mtcars)
tidy(mfx_logitreg, conf.int=T)

## Compare with naive model coefficients (not run)
# tidy(glm(am ~ cyl + hp + wt, family = binomial, data = mtcars), conf.int=T)

## One more example, this time using probit regression on the same data
mfx_probitreg <- probitmfx(am ~ cyl + hp + wt, atmean = T, data = mtcars)
tidy(mfx_probitreg, conf.int=T)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[mfx:betamfx]{mfx::betamfx()}}, \code{\link[mfx:logitmfx]{mfx::logitmfx()}}, \code{\link[mfx:negbinmfx]{mfx::negbinmfx()}}, \code{\link[mfx:poissonmfx]{mfx::poissonmfx()}}, \code{\link[mfx:probitmfx]{mfx::probitmfx()}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}
  \item{atmean}{were the marginal effects originally calculated as the partial effects for the average observation? If FALSE, then these were instead calculated as average partial effects.}

}
