% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-tidiers.R
\name{glance.gam}
\alias{glance.gam}
\title{Glance at a(n) gam object}
\usage{
\method{glance}{gam}(x, ...)
}
\arguments{
\item{x}{A \code{gam} object returned from a call to \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{df}{Degrees of freedom of the model.}
\item{logLik}{Log-likelihood of the model.}
\item{AIC}{Akaike's Information Criterion for the model.}
\item{BIC}{Bayesian Information Criterion for the model.}
\item{deviance}{Deviance of the model.}
\item{df.residual}{Residual degrees of freedom for the model.}
}
\description{
Description of what a \code{glance} method does goes here.
}
\details{
To glance \code{Gam} objects created by calls to \code{\link[gam:gam]{gam::gam()}}, see
\code{\link[=glance.Gam]{glance.Gam()}}.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[mgcv:gam]{mgcv::gam()}}, \code{\link[=glance.Gam]{glance.Gam()}}

Other mgcv tidiers: \code{\link{tidy.gam}}
}
\concept{mgcv tidiers}
