% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nls-tidiers}
\alias{augment.nls}
\alias{glance.nls}
\alias{nls-tidiers}
\alias{tidy.nls}
\title{Tidying methods for a nonlinear model}
\usage{
\method{tidy}{nls}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{augment}{nls}(x, data = NULL, ...)

\method{glance}{nls}(x, ...)
}
\arguments{
\item{x}{An object of class "nls"}

\item{data}{original data this was fitted on; if not given this will
attempt to be reconstructed from nls (may not be successful)}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{...}{extra arguments (not used)}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy} returns one row for each coefficient in the model,
with five columns:
  \item{term}{The term in the nonlinear model being estimated and tested}
  \item{estimate}{The estimated coefficient}
  \item{stderror}{The standard error from the linear model}
  \item{statistic}{t-statistic}
  \item{p.value}{two-sided p-value}

\code{augment} returns one row for each original observation,
with two columns added:
  \item{.fitted}{Fitted values of model}
  \item{.resid}{Residuals}

\code{glance} returns one row with the columns
  \item{sigma}{The square root of the estimated residual variance}
  \item{isConv}{Whether the fit successfully converged}
  \item{finTol}{The achieved convergence tolerance}
}
\description{
These methods tidy the coefficients of a nonlinear model into a summary,
augment the original data with information on the fitted values and residuals,
and construct a one-row glance of the model's statistics.
}
\seealso{
\code{\link{nls}} and \code{\link{summary.nls}}
}

