% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlslist_tidiers.R
\name{nlslist_tidiers}
\alias{nlslist_tidiers}
\alias{tidy.nlsList}
\title{Tidying methods for a list of nonlinear models}
\usage{
\method{tidy}{nlsList}(x, conf.int = FALSE, conf.level = 0.95,
  quick = FALSE, ...)
}
\arguments{
\item{x}{An object of class "nlsList"}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{quick}{whether to compute a smaller and faster version, containing
only the \code{term} and \code{estimate} columns.}

\item{data}{original data this was fitted on; if not given this will
attempt to be reconstructed from nlsList (may not be successful)}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy} returns one row for each coefficient in the model,
with five columns:
  \item{term}{The term in the nonlinear model being estimated and tested}
  \item{estimate}{The estimated coefficient}
  \item{std.error}{The standard error from the linear model}
  \item{statistic}{t-statistic}
  \item{p.value}{two-sided p-value}
}
\description{
These methods tidy the coefficients of a list of nonlinear models into a summary.
At the moment only the \code{tidy} method is available for this class of models.
This method applies to models fit with the \code{\link{nlsList}} function from the nlme package.
}
\examples{

chick <- as.data.frame(ChickWeight)

library(nlme)
nlsl1 <- nlsList(weight ~ a*Time^b | Diet, start=list(a=10, b=1), data=chick)

tidy(nlsl1)

}
\seealso{
\code{\link{nlsList}}, \code{\link{summary.nlsList}}.
}

