% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-polr-tidiers.R, R/ordinal-tidiers.R,
%   R/survey-tidiers.R
\name{tidy.polr}
\alias{tidy.polr}
\alias{glance.polr}
\alias{augment.polr}
\alias{ordinal_tidiers}
\alias{tidy.clm}
\alias{tidy.clmm}
\alias{glance.clm}
\alias{glance.clmm}
\alias{augment.clm}
\alias{tidy.svyolr}
\alias{glance.svyolr}
\title{Tidying methods for ordinal logistic regression models}
\usage{
\method{tidy}{polr}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)

\method{glance}{polr}(x, ...)

\method{augment}{polr}(x, data = model.frame(x), newdata = NULL,
  type.predict = c("probs", "class"), ...)

\method{tidy}{clm}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, conf.type = c("profile",
  "Wald"), ...)

\method{tidy}{clmm}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, conf.type = c("profile",
  "Wald"), ...)

\method{glance}{clm}(x, ...)

\method{glance}{clmm}(x, ...)

\method{augment}{clm}(x, data = model.frame(x), newdata = NULL,
  type.predict = c("prob", "class"), ...)

\method{tidy}{svyolr}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)

\method{glance}{svyolr}(x, ...)
}
\arguments{
\item{x}{a model of class \code{clm}, \code{clmm}, \code{polr} or \code{svyolr}}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{quick}{Logical indiciating if the only the \code{term} and \code{estimate}
columns should be returned. Often useful to avoid time consuming
covariance and standard error calculations. Defaults to \code{FALSE}.}

\item{...}{extra arguments}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors used to create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{type.predict}{type of prediction to compute for a CLM; passed on to
\code{\link[ordinal:predict.clm]{ordinal::predict.clm()}} or \code{predict.polr}}

\item{conf.type}{the type of confidence interval
(see \code{\link[ordinal:confint.clm]{ordinal::confint.clm()}})}
}
\value{
\code{tidy.clm}, \code{tidy.clmm}, \code{tidy.polr} and \code{tidy.svyolr}
return one row for each coefficient at each level of the response variable,
with six columns:
\item{term}{term in the model}
\item{estimate}{estimated coefficient}
\item{std.error}{standard error}
\item{statistic}{t-statistic}
\item{p.value}{two-sided p-value}
\item{coefficient_type}{type of coefficient, see \code{\link[ordinal:clm]{ordinal::clm()}}}

If \code{conf.int=TRUE}, it also includes columns for \code{conf.low} and

\code{glance.clm}, \code{glance.clmm}, \code{glance.polr} and \code{glance.svyolr}
return a one-row data.frame with the columns:
\item{edf}{the effective degrees of freedom}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{df.residual}{residual degrees of freedom}

\code{augment.clm} and \code{augment.polr} returns
one row for each observation, with additional columns added to
the original data:
\item{.fitted}{fitted values of model}
\item{.se.fit}{standard errors of fitted values}

\code{augment} is not supportted for \code{\link[ordinal:clmm]{ordinal::clmm()}}
and \code{\link[survey:svyolr]{survey::svyolr()}} models.

All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.
}
\description{
These methods tidy the coefficients of ordinal logistic regression
models generated by \code{\link[ordinal:clm]{ordinal::clm()}} or \code{\link[ordinal:clmm]{ordinal::clmm()}}
of the \code{ordinal} package, \code{\link[MASS:polr]{MASS::polr()}} of the \code{MASS}
packge, or \code{\link[survey:svyolr]{survey::svyolr()}} of the \code{survey} package.
}
\examples{
library(ordinal)
clm_mod <- clm(rating ~ temp * contact, data = wine)
tidy(clm_mod)
tidy(clm_mod, conf.int = TRUE)
tidy(clm_mod, conf.int = TRUE, conf.type = "Wald", exponentiate = TRUE)
glance(clm_mod)
augment(clm_mod)

clm_mod2 <- clm(rating ~ temp, nominal = ~ contact, data = wine)
tidy(clm_mod2)

clmm_mod <- clmm(rating ~ temp + contact + (1 | judge), data = wine)
tidy(clmm_mod)
glance(clmm_mod)

library(MASS)
polr_mod <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)
tidy(polr_mod, exponentiate = TRUE, conf.int = TRUE)
glance(polr_mod)
augment(polr_mod, type.predict = "class")
}
