% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm_tidiers.R
\name{process_ergm}
\alias{process_ergm}
\title{helper function to process a tidied ergm object}
\usage{
process_ergm(ret, x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE)
}
\arguments{
\item{ret}{data frame with a tidied version of a coefficient matrix}

\item{x}{an "ergm" object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for logistic regression)}
}
\description{
Optionally exponentiates the coefficients, and optionally adds
a confidence interval, to a tidied ergm object.
}
