% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_tidiers.R
\name{process_geeglm}
\alias{process_geeglm}
\title{helper function to process a tidied geeglm object}
\usage{
process_geeglm(ret, x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE)
}
\arguments{
\item{ret}{data frame with a tidied version of a coefficient matrix}

\item{x}{a "geeglm" object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for log distributions)}
}
\description{
Adds a confidence interval, and possibly exponentiates, a tidied
object.
}
