% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm_tidiers.R
\name{process_lm}
\alias{process_lm}
\title{helper function to process a tidied lm object}
\usage{
process_lm(ret, x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE)
}
\arguments{
\item{ret}{data frame with a tidied version of a coefficient matrix}

\item{x}{an "lm", "glm", "biglm", or "bigglm" object}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{exponentiate}{whether to exponentiate the coefficient estimates
and confidence intervals (typical for logistic regression)}
}
\description{
Adds a confidence interval, and possibly exponentiates, a tidied
object. Useful for operations shared between lm and biglm.
}

