% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rq_tidiers.R
\name{process_rq}
\alias{process_rq}
\title{Helper function for tidy.rq and tidy.rqs}
\usage{
process_rq(rq_obj, se.type = "rank", conf.int = TRUE, conf.level = 0.95,
  ...)
}
\arguments{
\item{rq_obj}{an object returned by \code{summary.rq} or \code{summary.rqs}}

\item{se.type}{type of standard errors used in \code{summary.rq} or \code{summary.rqs}}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level for confidence interval}

\item{\dots}{currently unused}
}
\description{
See documentation for \code{summary.rq} for complete description
of the options for \code{se.type}, \code{conf.int}, etc.
}

