% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcorr_tidiers.R
\name{rcorr_tidiers}
\alias{rcorr_tidiers}
\alias{tidy.rcorr}
\title{Tidying methods for rcorr objects}
\usage{
\method{tidy}{rcorr}(x, names = NULL, diagonal = FALSE, ...)
}
\arguments{
\item{x}{An object of class "rcorr"}

\item{names}{Optionally, the column names of the original
matrix, to use to fill in these columns.}

\item{diagonal}{Whether to include diagonal elements (where
\code{estimate} is 1 and \code{p.value} is NA), default FALSE}

\item{...}{extra arguments (not used)}
}
\value{
A data.frame with one row for each pairing
in the correlation matrix. Columns are:
  \item{column1}{The first column being described. A number, unless
  \code{names} is provided, in which case it will be the name}
  \item{column2}{The second column being described. A number, unless
  \code{names} is provided, in which case it will be the name}
  \item{estimate}{Estimate of Pearson's r or Spearman's rho}
  \item{n}{Number of observations used to compute the correlation}
  \item{p.value}{P-value of correlation}
}
\description{
Tidies a correlation matrix from the \code{rcorr} function in the
"Hmisc" package, including correlation estimates, p-values,
and the number of observations in each pairwise correlation.
Note that it returns these in "long", or "melted", format,
with one row for each pair of columns being compared.
}
\examples{
if (require("Hmisc", quietly = TRUE)) {
    mat <- replicate(100, rnorm(100))
    # add some NAs
    mat[sample(length(mat), 2000)] <- NA

    rc <- rcorr(mat)

    td <- tidy(rc)
    head(td)

    library(ggplot2)
    ggplot(td, aes(p.value)) +
        geom_histogram(binwidth = .1)

    ggplot(td, aes(estimate, p.value)) +
        geom_point() +
        scale_y_log10()
}
}

