% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/broom.R
=======
% Please edit documentation in R/broom-package.R, R/broom.R
>>>>>>> 3c922d507f7cc758a987a9ef44ae4267ac6ed583
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\seealso{
\code{\link[=augment.lm]{augment.lm()}}

\code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[=glance.lm]{glance.lm()}}
<<<<<<< HEAD
=======

\code{\link[=augment.lm]{augment.lm()}}

\code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[=glance.lm]{glance.lm()}}
>>>>>>> 3c922d507f7cc758a987a9ef44ae4267ac6ed583
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
<<<<<<< HEAD
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
=======
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}, \code{\link[generics]{tidy}}}
>>>>>>> 3c922d507f7cc758a987a9ef44ae4267ac6ed583
}}

