% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_tidiers.R
\name{rf_tidiers}
\alias{rf_tidiers}
\alias{tidy.randomForest}
\alias{augment.randomForest}
\alias{glance.randomForest}
\title{Tidying methods for a randomForest model}
\usage{
\method{tidy}{randomForest}(x, ...)

\method{augment}{randomForest}(x, data = NULL, ...)

\method{glance}{randomForest}(x, ...)
}
\arguments{
\item{x}{randomForest object}

\item{...}{Additional arguments (ignored)}

\item{data}{Model data for use by \code{\link{augment.randomForest}}.}
}
\value{
All tidying methods return a \code{data.frame} without rownames. The
  structure depends on the method chosen.

\code{tidy.randomForest} returns one row for each model term, with the following columns:
  \item{term}{The term in the randomForest model}
  \item{class_*}{One column for each model term; the relative importance of each term per class. Only present if the model was created with \code{importance = TRUE}}
  \item{MeanDecreaseAccuracy}{A measure of variable importance. See \code{\link[randomForest]{randomForest}} for more information. Only present if the model was created with \code{importance = TRUE}}
  \item{MeanDecreaseGini}{A measure of variable importance. See \code{\link[randomForest]{randomForest}} for more information.}
  \item{sd_*}{Sandard deviations for the preceding statistics. Only present if the model was created with \code{importance = TRUE}}

\code{augment.randomForest} returns the original data with additional columns:
  \item{.oob_times}{The number of trees for which the given case was "out of bag". See \code{\link[randomForest]{randomForest}} for more details.}
  \item{.fitted}{The fitted value or class.}
  \item{.li_*}{The casewise variable importance for each term. Only present if the model was created with \code{importance = TRUE}}
  In addition, \code{augment} returns additional columns for classification and usupervised trees:
  \item{.votes_*}{For each case, the voting results, with one column per class.}

\code{glance.randomForest} returns a one-row data.frame with
  the following columns:
  For regression trees, the following additional columns are present:
  \item{mse}{The average mean squared error across all trees.}
  \item{rsq}{The average pesudo-R-squared across all trees. See \code{\link[randomForest]{randomForest}} for more information.}
  For classification trees, the following columns are present for each class
  \item{*_precision}{}
  \item{*_recall}{}
  \item{*_accuracy}{}
  \item{*_f_measure}{}
}
\description{
These methods tidy the variable importance of a random forest model summary, 
augment the original data with information on the fitted
values/classifications and error, and construct a one-row glance of the
model's statistics.
}
