% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_tidiers.R
\name{rms_tidiers}
\alias{rms_tidiers}
\alias{tidy.cph}
\alias{tidy.lrm}
\alias{tidy.ols}
\alias{tidy.anova.rms}
\alias{tidy.summary.rms}
\alias{tidy.validate}
\alias{glance.rms}
\title{Tidiers for rms model objects}
\usage{
\method{tidy}{cph}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

\method{tidy}{lrm}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)

\method{tidy}{ols}(x, conf.int = TRUE, conf.level = TRUE,
  exponentiate = TRUE, ...)

\method{tidy}{anova.rms}(x, ...)

\method{tidy}{summary.rms}(x, ...)

\method{tidy}{validate}(x, concord = FALSE, ...)

\method{glance}{rms}(x, concord = FALSE, ...)
}
\arguments{
\item{x}{A model object that inherits class \code{rms}}

\item{conf.int}{Whether to return the confidence intervals in 
the tidy output}

\item{conf.level}{The confidence level for the confidence intervals}

\item{exponentiate}{Whether to exponentiate the estimate and
confidence intervals}

\item{...}{Extra arguments, not used}

\item{concord}{Whether the concordance index should be calculated,
if possible. For \code{glance.rms}, the concordance is only 
calculated if Somers' D (\code{Dxy}) is visible in the model's
\code{stats} element \emph{and} \code{C} is not already calculated.
For \code{tidy.validate}, the concordance index is calculated 
only when \code{Dxy} is calculated in \code{validate}.}
}
\description{
Tidiers for models built by the \code{rms} package, 
  including \code{ols}, \code{cph}, and \code{lrm}.  A \code{glance}
  method is provided for these models.  Additionally, 
  tidiers are made available for \code{anova.rms}, \code{summary.rms},
  and \code{validate} class objects derived from \code{rms} models.
}
\examples{
\dontrun{
library(rms)
DF <- data.frame(
  x1 = runif(200),
  x2 = runif(200),
  x3 = runif(200),
  x4 = runif(200)
)
DF$y <- DF$x1 + DF$x2 + rnorm(200)

fit <- ols(y ~ rcs(x1,4) + x2 + rcs(x3, 3) * x4, data = DF,
           x = TRUE, y = TRUE)

tidy(fit)
tidy(anova(fit))
glance(fit)

dd <- datadist(DF)
options(datadist = 'dd')
tidy(summary(fit))
tidy(validate(fit))


#* lrm Example
set.seed(17)
DF <- data.frame(
  age = rnorm(1000, 50, 10),
  cholesterol = rnorm(1000, 50, 10)
)
DF$ch <- Hmisc::cut2(DF$cholesterol, g=40, levels.mean=TRUE)

fit <- lrm(ch ~ age, data = DF, x = TRUE, y = TRUE)
tidy(fit)
tidy(anova(fit))
glance(fit)

dd <- datadist(DF)
options(datadist = 'dd')
tidy(summary(fit))
tidy(validate(fit))



#* cph Example
n <- 1000
set.seed(731)
DF <- data.frame(
    age = 50 + 12*rnorm(1000),
    sex = factor(sample(c('Male','Female'), 1000, 
                        rep=TRUE, prob=c(.6, .4))),
    cens = 15*runif(1000)
)

h <- .02*exp(.04*(DF$age-50)+.8*(DF$sex=='Female'))
DF$dt = -log(runif(1000) / h)
DF$e <- ifelse(DF$dt <= DF$cens,1,0)
DF$dt <- pmin(DF$dt, DF$cens)

fit <- cph(Surv(dt, e) ~ rcs(age,4) + sex, data = DF, 
           x=TRUE, y=TRUE, surv = TRUE)
tidy(fit)
tidy(anova(fit))
glance(fit)

dd <- datadist(DF)
options(datadist = 'dd')
tidy(summary(fit))
tidy(validate(fit))
}

}
