% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_tidiers.R
\name{robust_tidiers}
\alias{augment.glmRob}
\alias{augment.lmRob}
\alias{glance.glmRob}
\alias{glance.lmRob}
\alias{robust_tidiers}
\alias{tidy.glmRob}
\alias{tidy.lmRob}
\title{Tidiers for lmRob and glmRob objects}
\usage{
\method{tidy}{lmRob}(x, ...)

\method{augment}{lmRob}(x, ...)

\method{glance}{lmRob}(x, ...)

\method{tidy}{glmRob}(x, ...)

\method{augment}{glmRob}(x, ...)

\method{glance}{glmRob}(x, ...)
}
\description{
Tidying robust regression methods from the robust package. The tidy and augment
methods simply pass it on to the lm tidiers.
}
\examples{

if (requireNamespace("robust", quietly = TRUE)) {
  data(stack.dat)
  m <- lmRob(Loss ~ ., data = stack.dat)
  
  tidy(m)
  augment(m)
  glance(m)
}

}

