% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_tidiers.R
\name{sparse_tidiers}
\alias{sparse_tidiers}
\alias{tidy.dgCMatrix}
\alias{tidy.dgTMatrix}
\alias{tidy.sparseMatrix}
\title{Tidy a sparseMatrix object from the Matrix package}
\usage{
\method{tidy}{dgTMatrix}(x, ...)

\method{tidy}{dgCMatrix}(x, ...)

\method{tidy}{sparseMatrix}(x, ...)
}
\arguments{
\item{x}{A Matrix object}

\item{...}{Extra arguments, not used}
}
\description{
Tidy a sparseMatrix object from the Matrix package into
a three-column data frame, row, column, and value (with
zeros missing). If there are row names or column names,
use those, otherwise use indices
}

