% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedglm-tidiers.R
\name{speedlm_tidiers}
\alias{speedlm_tidiers}
\alias{tidy.speedlm}
\alias{glance.speedlm}
\alias{augment.speedlm}
\title{Tidying methods for a speedlm model}
\usage{
\method{tidy}{speedlm}(x, ...)

\method{glance}{speedlm}(x, ...)

\method{augment}{speedlm}(x, data = stats::model.frame(x), newdata = data,
  ...)
}
\arguments{
\item{x}{speedlm object}

\item{...}{extra arguments (not used)}

\item{data}{data frame to augment}

\item{newdata}{new data to use for predictions, optional}
}
\value{
All tidying methods return a data.frame without rownames, whose
structure depends on the method chosen.

\code{tidy.speedlm} returns the tidied output of the
lm with one row for each term in the formula.
The columns match those in \code{\link[=tidy.lm]{tidy.lm()}}.

\code{glance.speedlm} returns a one-row data.frame with the columns
\item{r.squared}{The percent of variance explained by the model}
\item{adj.r.squared}{r.squared adjusted based on the degrees of freedom}
\item{statistic}{F-statistic}
\item{p.value}{p-value from the F test, describing whether the full
regression is significant}
\item{df}{Degrees of freedom used by the coefficients}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}

\code{augment.speedlm} returns  one row for each observation, with just one column added:
\item{.fitted}{Fitted values of model}
}
\description{
These methods tidy the coefficients of a "speedlm" object
into a summary, augment the original data with information on the
fitted values and residuals, and construct a one-row glance of the model's
statistics.
}
\examples{

if (require("speedglm", quietly = TRUE)) {
    mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars)
    tidy(mod)
    glance(mod)
    augment(mod)
}

}
\seealso{
\code{\link[=tidy.lm]{tidy.lm()}}, \code{\link[=tidy.biglm]{tidy.biglm()}}
}
