% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_tidiers.R
\name{survdiff_tidiers}
\alias{glance.survdiff}
\alias{survdiff_tidiers}
\alias{tidy.survdiff}
\title{Tidiers for Tests of Differences between Survival Curves}
\usage{
\method{tidy}{survdiff}(x, strata = FALSE, ...)

\method{glance}{survdiff}(x, ...)
}
\arguments{
\item{x}{a "survdiff" object}

\item{strata}{logical, whether to include strata in the output}

\item{other}{arguments passed to/from other methods, currently ignored#'}
}
\value{
All tidying methods return a data.frame without rownames, whose
structure depends on the method chosen.

TODO Describe output columns

TODO describe output columns
}
\description{
Tidiers for Tests of Differences between Survival Curves
}

