% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo-tidiers.R
\name{tidy.apollo}
\alias{tidy.apollo}
\alias{apollo_tidiers}
\title{Tidying methods for Apollo model outputs}
\usage{
\method{tidy}{apollo}(
  x,
  conf.int = FALSE,
  se.type = c("default", "robust"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{an object returned from \code{\link[apollo:apollo_estimate]{apollo::apollo_estimate()}}}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
These functions tidy the coefficients of abstract discrete choice models
estimated with the \code{apollo} package.
}
\details{
As of now, the objects returned from apollo are instances of
class \code{maxLik}.  This requires use of the method call, which might disrupt
some downstream operations. It might be appropriate to manually change the
model object class to \code{apollo}.
}
\examples{
\dontrun{
  apollo_mnl
  tidy.apollo(apollo_mnl) 
  tidy.apollo(apollo_mnl, se.type = "robust") 
  tidy.apollo(apollo_mnl, conf.int = TRUE) 

  glance.apollo(apollo_mnl)
}

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[apollo:apollo_estimate]{apollo::apollo_estimate()}}
}
\concept{apollo tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}

}
