% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.data.frame}
\alias{tidy.data.frame}
\title{tidy on a data frame}
\usage{
\method{tidy}{data.frame}(x, ...)
}
\arguments{
\item{x}{A data.frame}

\item{...}{extra arguments (not used)}
}
\value{
The same object given to the function
}
\description{
tidy on a data frame input returns the same object. There is probably
never a reason to do this, unless a function may return a data
frame and may return an untidy output.
}

