% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tidy.default}
\alias{tidy.default}
\title{Default tidying method}
\usage{
\method{tidy}{default}(x, ...)
}
\arguments{
\item{x}{an object to be tidied}

\item{...}{extra arguments (not used)}
}
\value{
A data frame, from \code{as.data.frame} applied to the input x.
}
\description{
By default, tidy uses \code{as.data.frame} to convert its output. This is
dangerous, as it may fail with an uninformative error message.
Generally tidy is intended to be used on structured model objects
such as lm or htest for which a specific S3 object exists.
}
\details{
If you know that you want to use \code{as.data.frame} on your untidy
object, just use it directly.
}

