% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiR-tidiers.R
\name{tidy.epi.2by2}
\alias{tidy.epi.2by2}
\alias{epiR_tidiers}
\title{Tidy a(n) epi.2by2 object}
\usage{
\method{tidy}{epi.2by2}(x, ...)
}
\arguments{
\item{x}{A \code{epi.2by2} object produced by a call to \code{\link[epiR:epi.2by2]{epiR::epi.2by2()}}}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
The tibble has a column for each of the measures of association or tests contained in \code{massoc} when
\code{\link[epiR:epi.2by2]{epiR::epi.2by2()}} is called.
}
\examples{

library(epiR)

dat <- matrix(c(13,2163,5,3349), nrow = 2, byrow = TRUE)
rownames(dat) <- c("DF+", "DF-")
colnames(dat) <- c("FUS+", "FUS-")
birthwt$low <- factor(birthwt$low, levels = c(1,0))
birthwt$smoke <- factor(birthwt$smoke, levels = c(1,0))
birthwt$race <- factor(birthwt$race, levels = c(1,2,3))
tab1 <- table(birthwt$smoke, birthwt$low, dnn = c("Smoke", "Low BW"))

tidy(tab1)
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[epiR:epi.2by2]{epiR::epi.2by2()}}
}
\concept{epiR tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{df}{Degrees of freedom used by this term in the model.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{parameter}{The parameter being modeled.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{estimate}{Estimated measure of association}

}
