% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.ftable}
\alias{tidy.ftable}
\title{tidy an ftable object}
\usage{
\method{tidy}{ftable}(x, ...)
}
\arguments{
\item{x}{An object of class "ftable"}

\item{...}{Extra arguments (not used)}
}
\value{
A data.frame with a column for each variable that has been counted,
named \code{Var1}, \code{Var2}, etc, then
}
\description{
A table contains a contingency table that
}

