% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multcomp-tidiers.R
\name{tidy.glht}
\alias{tidy.glht}
\alias{multcomp_tidiers}
\title{Tidy a(n) glht object}
\usage{
\method{tidy}{glht}(x, ...)
}
\arguments{
\item{x}{A \code{glht} object returned by \code{\link[multcomp:glht]{multcomp::glht()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Description of what a \code{tidy} method does goes here.
}
\examples{

if (require("multcomp") && require("ggplot2")) {

    library(multcomp)
    library(ggplot2)
    
    amod <- aov(breaks ~ wool + tension, data = warpbreaks)
    wht <- glht(amod, linfct = mcp(tension = "Tukey"))

    tidy(wht)
    ggplot(wht, aes(lhs, estimate)) + geom_point()

    CI <- confint(wht)
    tidy(CI)
    ggplot(CI, aes(lhs, estimate, ymin = lwr, ymax = upr)) +
       geom_pointrange()

    tidy(summary(wht))
    ggplot(mapping = aes(lhs, estimate)) +
       geom_linerange(aes(ymin = lwr, ymax = upr), data = CI) +
       geom_point(aes(size = p), data = summary(wht)) +
       scale_size(trans = "reverse")

    cld <- cld(wht)
    tidy(cld)
}

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[multcomp:glht]{multcomp::glht()}}

Other multcomp tidiers: \code{\link{tidy.cld}},
  \code{\link{tidy.confint.glht}},
  \code{\link{tidy.summary.glht}}
}
