% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-glm-tidiers.R
\name{tidy.glm}
\alias{tidy.glm}
\title{Tidy a(n) glm object}
\usage{
\method{tidy}{glm}(x, ...)
}
\arguments{
\item{x}{A \code{glm} object returned from \code{\link[stats:glm]{stats::glm()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{}{}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[stats:glm]{stats::glm()}}

Other lm tidiers: \code{\link{augment.glm}},
  \code{\link{glance.glm}}
}
\concept{lm tidiers}
