% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tidiers.R
\name{tidy.Kendall}
\alias{tidy.Kendall}
\title{Tidying methods for Kendall objects}
\usage{
\method{tidy}{Kendall}(x)
}
\arguments{
\item{x}{Fitted object from the Kendall package.}
}
\value{
\code{tidy()} returns a \code{\link[tibble]{tibble}} with one row and columns:

\item{statistic}{Kendall's tau statistic.}
\item{p.value}{two-sided p-value.}
\item{kendall_score}{Kendall score.}
\item{denominator}{The denominator, which is tau=kendall_score/denominator.}
\item{var_kendall_score}{Variance of the kendall_score.}
}
\description{
Tidies the result of an Kendall rank correlation. Works on all functions from the Kendall package.
Only a \code{tidy()} method is provided, not an \code{augment()} or \code{glance()} method.
}
\examples{

\dontrun{
library(Kendall)

### For Kendall ####
A <- c(2.5,2.5,2.5,2.5,5,6.5,6.5,10,10,10,10,10,14,14,14,16,17)
B <- c(1,1,1,1,2,1,1,2,1,1,1,1,1,1,2,2,2)
f_res <- Kendall(A,B)

tidy(f_res)

### For MannKendall ####
s_res <- MannKendall(B)

tidy(s_res)

### For SeasonalMannKendall ####
t_res <- SeasonalMannKendall(ts(A))

tidy(t_res)
}
}
\seealso{
\code{\link[Kendall]{Kendall}} for details on the meaning of each column and implementation.
}
