% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-lm-tidiers.R
\name{tidy.lm.beta}
\alias{tidy.lm.beta}
\alias{tidy.summary.lm.beta}
\title{TODO: documentation}
\usage{
\method{tidy}{lm.beta}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, ...)

\method{tidy}{summary.lm.beta}(x, ...)
}
\description{
TODO: documentation

TODO: document
}
\examples{
CarReg = lm(formula = speed ~., data = cars)

# Standardize coefficients
library(lm.beta)
CarRegS = lm.beta::lm.beta(CarReg)

# Table format
tidy(CarRegS)

}
\seealso{
Other lm tidiers: \code{\link{augment.glm}},
  \code{\link{augment.lm}}, \code{\link{glance.glm}},
  \code{\link{glance.lm}}, \code{\link{tidy.glm}},
  \code{\link{tidy.lm}}
}
\concept{lm tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}

}
