\name{tidy.lm}
\alias{tidy.lm}
\title{tidy a linear model by returning a data frame version of the coefficients
table}
\usage{
\method{tidy}{lm}(x, ...)
}
\arguments{
  \item{x}{An object of class "lm"}

  \item{...}{extra arguments (not used)}
}
\value{
a data.frame with four columns:
}
\description{
tidy a linear model by returning a data frame version of
the coefficients table
}

