% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-lmrob-tidiers.R
\name{tidy.lmRob}
\alias{tidy.lmRob}
\alias{robust_tidiers}
\title{Tidy a(n) lmRob object}
\usage{
\method{tidy}{lmRob}(x, ...)
}
\arguments{
\item{x}{A \code{lmRob} object returned from \code{\link[robust:lmRob]{robust::lmRob()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{}{}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}.
}
\examples{

library(robust)
m <- lmRob(mpg ~ wt, data = mtcars)

tidy(m)
augment(m)
glance(m)

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
glance(gm)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[robust:lmRob]{robust::lmRob()}}

Other robust tidiers: \code{\link{augment.glmRob}},
  \code{\link{augment.lmRob}}, \code{\link{glance.glmRob}},
  \code{\link{glance.lmRob}}, \code{\link{tidy.glmRob}}
}
\concept{robust tidiers}
