% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistf-tidiers.R
\name{tidy.logistf}
\alias{tidy.logistf}
\alias{logistf_tidiers}
\title{Tidy a(n) logistf object}
\usage{
\method{tidy}{logistf}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A \code{logistf} object returned from \code{\link[logistf:logistf]{logistf::logistf()}}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.level = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies across models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
The \code{logistf} package implements "logistic regression model using
Firth's bias reduction method."
}
\examples{
library(logistf)

dat <- data.frame(time = 1:12, 
                  event = c(0,1,0,1,1,1,1,1,1,1,1,1))
mod <- logistf::logistf(event ~time, dat)
tidy(mod)
glance(mod)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[logistf:logistf]{logistf::logistf()}}
}
\concept{logistf tdiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}

}
