% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-tidiers.R
\name{tidy.lsmobj}
\alias{tidy.lsmobj}
\alias{emmeans_tidiers}
\title{Tidy a(n) lsmobj object}
\usage{
\method{tidy}{lsmobj}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An \code{lsmobj} object.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments passed to \code{emmeans::summary.emmGrid()} or
\code{lsmeans::summary.ref.grid()}. \strong{Cautionary note}: misspecified arguments
may be silently ignored!}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies across models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
Returns a data frame with one observation for each estimated marginal
mean, and one column for each combination of factors. When the input is a
contrast, each row will contain one estimated contrast.

There are a large number of arguments that can be
passed on to \code{emmeans::summary.emmGrid()} or \code{lsmeans::summary.ref.grid()}.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{emmeans::ref_grid()}, \code{emmeans::emmeans()},
\code{emmeans::contrast()}

Other emmeans tidiers: 
\code{\link{tidy.emmGrid}()},
\code{\link{tidy.ref.grid}()},
\code{\link{tidy.summary_emm}()}
}
\concept{emmeans tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{contrast}{Levels being compared.}
  \item{df}{Degrees of freedom used by this term in the model.}
  \item{null.value}{Value to which the estimate is compared.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{std.error}{The standard error of the regression term.}
  \item{estimate}{Expected marginal mean}
  \item{statistic}{T-ratio statistic}

}
