% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.manova}
\alias{tidy.manova}
\title{tidy a MANOVA object}
\usage{
\method{tidy}{manova}(x, ...)
}
\arguments{
\item{x}{object of class "manova"}

\item{...}{additional arguments (not used)}
}
\description{
Constructs a data frame with one row for each of the terms in the model,
containing the information from \link{summary.manova}.
}

