% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit-tidiers.R
\name{tidy.mlogit}
\alias{tidy.mlogit}
\alias{mlogit_tidiers}
\title{Tidying methods for multinomial logistic regression models}
\usage{
tidy.mlogit(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, ...)
}
\arguments{
\item{x}{A \code{mlogit} object returned from \code{\link[mlogit:mlogit]{mlogit::mlogit()}}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
These methods tidy the coefficients of multinomial logit regression
models generated by \code{mlogit} of the \code{mlogit} package.
}
\examples{

library(mlogit)

data("Fishing", package = "mlogit")
Fish <- mlogit.data(Fishing, varying = c(2:9), shape = "wide", choice = "mode")
fish.mnl <- mlogit(mode ~ price + catch, data = Fish)

tidy(fish.mnl)
tidy(fish.mnl, conf.int = 0.9)
glance(fish.mnl)
augment(fish.mnl)


}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[mlogit:mlogit]{mlogit::mlogit()}}

Other mlogit tidiers: \code{\link{augment.mlogit}},
  \code{\link{glance.mlogit}}
}
\concept{mlogit tidiers}
