% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet-tidiers.R
\name{tidy.multinom}
\alias{tidy.multinom}
\alias{multinom_tidiers}
\alias{nnet_tidiers}
\title{Tidying methods for multinomial logistic regression models}
\usage{
\method{tidy}{multinom}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = TRUE, ...)
}
\arguments{
\item{x}{A \code{multinom} object returned from \code{\link[nnet:multinom]{nnet::multinom()}}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
\code{tidy.multinom} returns one row for each coefficient at each
level of the response variable, with six columns:
\item{y.value}{The response level}
\item{term}{The term in the model being estimated and tested}
\item{estimate}{The estimated coefficient}
\item{std.error}{The standard error from the linear model}
\item{statistic}{Wald z-statistic}
\item{p.value}{two-sided p-value}

If \code{conf.int = TRUE}, also includes columns for \code{conf.low} and
\code{conf.high}.
}
\description{
These methods tidy the coefficients of multinomial logistic regression
models generated by \code{multinom} of the \code{nnet} package.
}
\examples{

if (require(nnet) & require(MASS)){
  library(nnet)
  library(MASS)
  
  example(birthwt)
  bwt.mu <- multinom(low ~ ., bwt)
  tidy(bwt.mu)
  glance(bwt.mu)

  #* This model is a truly terrible model
  #* but it should show you what the output looks
  #* like in a multinomial logistic regression

  fit.gear <- multinom(gear ~ mpg + factor(am), data = mtcars)
  tidy(fit.gear)
  glance(fit.gear)
}

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[nnet:multinom]{nnet::multinom()}}

Other multinom tidiers: \code{\link{glance.multinom}}
}
\concept{multinom tidiers}
