% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcutt-tidiers.R
\name{tidy.orcutt}
\alias{tidy.orcutt}
\alias{orcutt_tidiers}
\title{Tidy a(n) orcutt object}
\usage{
\method{tidy}{orcutt}(x, ...)
}
\arguments{
\item{x}{An \code{orcutt} object returned from \code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{}{}
}}
}
\value{
Something goes here

\code{tidy} returns the same information as \code{\link[=tidy.lm]{tidy.lm()}},
though without confidence interval options.
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\examples{

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

if (require("orcutt", quietly = TRUE)) {
  co <- cochrane.orcutt(reg)
  co

  tidy(co)
  glance(co)
}

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}

Other orcutt tidiers: \code{\link{glance.orcutt}}
}
\concept{orcutt tidiers}
