% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-tidiers.R
\name{tidy.pam}
\alias{tidy.pam}
\alias{pam_tidiers}
\title{These methods summarize the results of Partitioning Around Medoids (PAM) clustering into three
tidy forms. \code{tidy} describes the medoid and size of each cluster,
\code{augment} adds the cluster assignments to the original data, and
\code{glance} summarizes the average silhouette with of the clustering.}
\usage{
\method{tidy}{pam}(x, col.names = paste0("x", 1:ncol(x$medoids)), ...)
}
\arguments{
\item{x}{An \code{pam} object returned from \code{\link[cluster:pam]{cluster::pam()}}}

\item{col.names}{Dimension names. Defaults to the names of the variables in x.  Set to NULL to get names \code{x1, x2, ...}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
These methods summarize the results of Partitioning Around Medoids (PAM) clustering into three
tidy forms. \code{tidy} describes the medoid and size of each cluster,
\code{augment} adds the cluster assignments to the original data, and
\code{glance} summarizes the average silhouette with of the clustering.

Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
For examples, see the pam vignette.
}
\examples{

library(dplyr)
library(ggplot2)
library(cluster)

x<- iris \%>\%select(-Species)
p<-pam(x, k = 3)

tidy(p)
glance(p)
augment(p,x)
ggplot(augment(p,x), aes(Sepal.Length, Sepal.Width)) + geom_point(aes(color = .cluster)) +geom_text(aes(label = cluster), data = tidy(p), size = 10)
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[cluster:pam]{cluster::pam()}}

Other pam tidiers: \code{\link{augment.pam}},
  \code{\link{glance.pam}}
}
