% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aTSA.tidiers.R
\name{tidy}
\alias{tidy}
\title{Tidying methods for Dickey-Fuller unit root test}
\usage{
tidy(x, alpha = 0.05)
}
\arguments{
\item{x}{An object of object returned from \code{\link[aTSA:adf.test]{aTSA::adf.test()}} or \code{\link[aTSA:arch.test]{aTSA::arch.test()}}.}
}
\description{
These methods tidy test of unit root for time series
models generated by functions of the \code{aTSA} package.
}
\examples{
library(aTSA) 
library(tidyverse)
x <- arima.sim(list(order = c(1,0,0),ar = 0.2),n = 100)
A =  adf.test(x)
Y = 
adf.test(x)

Y = ts(AirPassengers)
adf.test(Y)



}
