% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustbase-lmrob-tidiers.R
\name{tidy.lmrob}
\alias{tidy.lmrob}
\alias{robustbase_tidiers}
\title{Tidy a(n) lmrob object}
\usage{
\method{tidy}{lmrob}(x, ...)
}
\arguments{
\item{x}{A \code{lmrob} object returned from \code{\link[robustbase:lmrob]{robustbase::lmrob()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
This method wraps \code{\link[=tidy.lm]{tidy.lm()}}.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}. For tidiers for robust models from the \pkg{robust} package
see \code{\link[=tidy.lmRob]{tidy.lmRob()}}.
}
\examples{

# From the robustbase::lmrob examples:
data(coleman)
set.seed(0)

m <- robustbase::lmrob(Y ~ ., data=coleman)
tidy(m)
augment(m)
glance(m)

TODO: glm
# From the robustbase::glmrob examples:
data(carrots)
Rfit <- glmrob(cbind(success, total-success) ~ logdose + block,
                family = binomial, data = carrots, method= "Mqle",
                control= glmrobMqle.control(tcc=1.2))
tidy(Rfit)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[robustbase:lmrob]{robustbase::lmrob()}}

Other robustbase tidiers: \code{\link{augment.lmrob}},
  \code{\link{glance.glmrob}}, \code{\link{glance.lmrob}},
  \code{\link{tidy.glmrob}}
}
\concept{robustbase tidiers}
