% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tidy.spec}
\alias{tidy.spec}
\title{tidy a spec objet}
\usage{
\method{tidy}{spec}(x, ...)
}
\arguments{
\item{x}{an object of class "spec"}

\item{...}{extra arguments (not used)}
}
\value{
a data frame with "freq" and "spec" columns
}
\description{
Given a "spec" object, which shows a spectrum across a range of frequencies,
returns a tidy data frame with two columns: "freq" and "spec"
}
\examples{
spc <- spectrum(lh)
tidy(spc)

library(ggplot2)
ggplot(tidy(spc), aes(freq, spec)) + geom_line()
}

