% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedglm-tidiers.R
\name{tidy.speedlm}
\alias{tidy.speedlm}
\alias{speedlm_tidiers}
\alias{speedglm_tidiers}
\title{Tidy a(n) speedlm object}
\usage{
\method{tidy}{speedlm}(x, ...)
}
\arguments{
\item{x}{A \code{speedlm} object returned from \code{\link[speedglm:speedlm]{speedglm::speedlm()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{}{}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\examples{

mod <- speedglm::speedlm(mpg ~ wt + qsec, data = mtcars)

tidy(mod)
glance(mod)
augment(mod)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[speedglm:speedlm]{speedglm::speedlm()}}

Other speedlm tidiers: \code{\link{augment.speedlm}},
  \code{\link{glance.speedlm}}
}
\concept{speedlm tidiers}
