% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rstan_tidiers.R
\name{tidy.stanfit}
\alias{tidy.stanfit}
\title{Tidying methods for Stan fits}
\usage{
\method{tidy}{stanfit}(x, pars, pt.method = "mean", conf.int = FALSE,
  conf.level = 0.95, conf.method = "quantile", ...)
}
\arguments{
\item{x}{an object of class \sQuote{"stanfit"}}

\item{pars}{(character) specification of which parameters to include}

\item{pt.method}{method for computing point estimate ("mean" or "median")}

\item{conf.int}{(logical) include confidence interval?}

\item{conf.level}{probability level for CI}

\item{conf.method}{method for computing confidence intervals ("quantile" or "HPDinterval")}

\item{...}{unused}
}
\description{
Tidying methods for Stan fits
}

