% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-anova-tidiers.R
\name{tidy.summary.manova}
\alias{tidy.summary.manova}
\title{Tidy a(n) summary.manova object}
\usage{
\method{tidy}{summary.manova}(x, ...)
}
\arguments{
\item{x}{A \code{summary.manova} object return from \code{\link[stats:summary.manova]{stats::summary.manova()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
Depending on which test statistic was calculated when the object
was created, only one of \code{pillai}, \code{wilks}, \code{hl} or \code{roy} is included.
}
\examples{

npk2 <- within(npk, foo <- rnorm(24))

m <- summary(
  manova(cbind(yield, foo) ~ block + N * P * K, npk2),
  test = "Wilks"
)

tidy(m)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[stats:summary.manova]{stats::summary.manova()}}

Other anova tidiers: 
\code{\link{glance.aov}()},
\code{\link{tidy.TukeyHSD}()},
\code{\link{tidy.anova}()},
\code{\link{tidy.aovlist}()},
\code{\link{tidy.aov}()},
\code{\link{tidy.manova}()}
}
\concept{anova tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{den.df}{Degrees of freedom of the denominator}
  \item{num.df}{Degrees of freedom}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{term}{The name of the regression term.}
  \item{pillai}{Pillai's trace.}
  \item{wilks}{Wilk's lambda.}
  \item{hl}{Hotelling-Lawley trace.}
  \item{roy}{Roy's greatest root.}

}
