% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans-tidiers.R
\name{tidy.summary_emm}
\alias{tidy.summary_emm}
\title{Tidy a(n) summary_emm object}
\usage{
\method{tidy}{summary_emm}(x, null.value = NULL, ...)
}
\arguments{
\item{x}{A \code{summary_emm} object.}

\item{null.value}{Value to which estimate is compared.}

\item{...}{Additional arguments passed to \code{emmeans::summary.emmGrid()} or
\code{lsmeans::summary.ref.grid()}. \strong{Cautionary note}: misspecified arguments
may be silently ignored!}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies across models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
Returns a data frame with one observation for each estimated marginal
mean, and one column for each combination of factors. When the input is a
contrast, each row will contain one estimated contrast.

There are a large number of arguments that can be
passed on to \code{emmeans::summary.emmGrid()} or \code{lsmeans::summary.ref.grid()}.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{emmeans::ref_grid()}, \code{emmeans::emmeans()},
\code{emmeans::contrast()}

Other emmeans tidiers: 
\code{\link{tidy.emmGrid}()},
\code{\link{tidy.lsmobj}()},
\code{\link{tidy.ref.grid}()}
}
\concept{emmeans tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{contrast}{Levels being compared.}
  \item{den.df}{Degrees of freedom of the denominator.}
  \item{df}{Degrees of freedom used by this term in the model.}
  \item{null.value}{Value to which the estimate is compared.}
  \item{num.df}{Degrees of freedom.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{std.error}{The standard error of the regression term.}
  \item{level1}{One level of the factor being contrasted}
  \item{level2}{The other level of the factor being contrasted}
  \item{term}{Model term in joint tests}
  \item{estimate}{Expected marginal mean}
  \item{statistic}{T-ratio statistic or F-ratio statistic}

}
