% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.table}
\alias{tidy.table}
\title{tidy a table object}
\usage{
\method{tidy}{table}(x, ...)
}
\arguments{
\item{x}{An object of class "table"}

\item{...}{Extra arguments (not used)}
}
\value{
A data.frame with a column for each variable that has been counted,
named \code{Var1}, \code{Var2}, etc, then
}
\description{
A table, typically created by the \link{table} function, contains a contingency
table of frequencies across multiple vectors.
}

