% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.TukeyHSD}
\alias{tidy.TukeyHSD}
\title{tidy a TukeyHSD object}
\usage{
\method{tidy}{TukeyHSD}(x, ...)
}
\arguments{
\item{x}{object of class "TukeyHSD"}

\item{...}{additional arguments (not used)}
}
\description{
Returns a data.frame with one row for each pairwise comparison
}

