% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscl-tidiers.R
\name{tidy_zip}
\alias{tidy_zip}
\alias{tidy_zeroinfl}
\alias{tidy_hurdle}
\title{Turn zero-inflated model results into a tidy tibble}
\usage{
tidy_zeroinfl(x, type = "count", ...)

tidy_hurdle(x, type = "count", ...)
}
\arguments{
\item{x}{A \code{hurdle} or \code{zeroinfl} model object.}

\item{type}{A character string for which model coefficients to return:
"all", "count", or "zero".}

\item{...}{Not currently used.}
}
\value{
A tibble
}
\description{
Turn zero-inflated model results into a tidy tibble
}
