% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-tidiers.R
\name{xgb_feature_importance}
\alias{xgb_feature_importance}
\title{Extract xgboost importance table}
\usage{
xgb_feature_importance(model, ...)
}
\arguments{
\item{model}{an xgb.Booster object}

\item{...}{arguments to be passed to xgboost::xgb.importance()}
}
\value{
tibble::tibble() containing gain, cover, frequency for each feature.
}
\description{
Extract xgboost importance table
}
