% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost-tidiers.R
\name{xgb_trees}
\alias{xgb_trees}
\title{Extract xgboost tree structure}
\usage{
xgb_trees(model, ...)
}
\arguments{
\item{model}{an xgb.Booster object}

\item{...}{arguments to be passed to xgboost::xgb.model.dt.tree()}
}
\value{
tibble::tibble() containing the tree structure - one row per split decision
}
\description{
Extract xgboost tree structure
}
