% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-htest-tidiers.R
\name{augment.htest}
\alias{augment.htest}
\title{Augment data with information from a(n) htest object}
\usage{
\method{augment}{htest}(x, ...)
}
\value{
Errors unless \code{x} is a chi-squared test. If \code{x} is a chi-squared
test, for each cell of the tested table returns columns:

\item{.observed}{Observed count}
\item{.prop}{Proportion of the total}
\item{.row.prop}{Row proportion (2 dimensions table only)}
\item{.col.prop}{Column proportion (2 dimensions table only)}
\item{.expected}{Expected count under the null hypothesis}
\item{.residuals}{Pearson residual}
\item{.stdres}{Standardized residual}
}
\description{
Description of what a \code{augment} method does goes here.
}
\details{
See \code{\link[stats:chisq.test]{stats::chisq.test()}} for more details on
how residuals are computed.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[stats:chisq.test]{stats::chisq.test()}}

Other htest tidiers: \code{\link{htest_tidiers}},
  \code{\link{tidy.pairwise.htest}},
  \code{\link{tidy.power.htest}}
}
\concept{htest tidiers}
