% Generated by roxygen2 (4.0.1): do not edit by hand
\name{augment.kmeans}
\alias{augment.kmeans}
\title{augment a kmeans object}
\usage{
\method{augment}{kmeans}(model, data, ...)
}
\arguments{
\item{model}{an object of class "kmeans"}

\item{data}{the original data that was clustered (required)}

\item{...}{extra arguments (not used)}
}
\value{
The original data with one extra column:
  \item{.cluster}{The cluster assigned by the k-means algorithm}
}
\description{
Constructs a data.frame containing one row for each initial point, with
the cluster assignment added.
}

