% Generated by roxygen2 (4.0.1): do not edit by hand
\name{augment.nls}
\alias{augment.nls}
\title{augment method for nls objects}
\usage{
\method{augment}{nls}(model, data = NULL, ...)
}
\arguments{
\item{model}{a fitted nls object}

\item{data}{original data this was fitted on; if not given this will
attempt to be reconstructed from nls (may not be successful)}

\item{...}{not used by this method}
}
\value{
The original data with extra columns:
  \item{.fitted}{Fitted values of model}
  \item{.resid}{Residuals}

The original data
}
\description{
This combines the original data given to smooth.spline with the
fit and residuals
}

