% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Augment data according to a tidied model}
\usage{
augment(x, data, ...)
}
\arguments{
\item{x}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed (when possible this is extracted
from the model)}

\item{...}{other arguments passed to methods}
}
\description{
Given an R statistical model or other non-tidy object, add columns to the
original dataset such as predictions, residuals and cluster assignments.
}
\details{
This generic originated in the ggplot2 package, where it was called
"fortify."
}
\seealso{
\code{\link{augment.lm}}
}

