% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-rlm-tidiers.R
\name{augment.rlm}
\alias{augment.rlm}
\title{Tidy a(n) rlm object}
\usage{
\method{augment}{rlm}(x, ...)
}
\arguments{
\item{x}{An \code{rlm} object returned by \code{\link[MASS:rlm]{MASS::rlm()}}.}

\item{...}{Arguments passed on to \code{augment.lm}
\describe{
  \item{x}{lm object}
  \item{data}{Original data, defaults to the extracting it from the model}
  \item{newdata}{If provided, performs predictions on the new data}
  \item{type.predict}{Type of prediction to compute for a GLM; passed on to
\code{\link[=predict.glm]{predict.glm()}}}
  \item{type.residuals}{Type of residuals to compute for a GLM; passed on to
\code{\link[=residuals.glm]{residuals.glm()}}}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{augment} method does goes here. Wraps
\code{\link[=augment.lm]{augment.lm()}}. Details on this. Document params other than \code{x}.
}
\details{
For tidiers for models from the \pkg{robust} package see
\code{\link[=tidy.lmRob]{tidy.lmRob()}} and \code{\link[=tidy.glmRob]{tidy.glmRob()}}.
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[=augment.lm]{augment.lm()}}

\code{\link[MASS:rlm]{MASS::rlm()}}

Other rlm tidiers: \code{\link{glance.rlm}},
  \code{\link{tidy.rlm}}
}
