% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{augment_columns}
\alias{augment_columns}
\title{add fitted values, residuals, and other common outputs to
an augment call}
\usage{
augment_columns(
  x,
  data,
  newdata = NULL,
  type,
  type.predict = type,
  type.residuals = type,
  se.fit = TRUE,
  ...
)
}
\arguments{
\item{x}{a model}

\item{data}{original data onto which columns should be added}

\item{newdata}{new data to predict on, optional}

\item{type}{Type of prediction and residuals to compute}

\item{type.predict}{Type of prediction to compute; by default
same as \code{type}}

\item{type.residuals}{Type of residuals to compute; by default
same as \code{type}}

\item{se.fit}{Value to pass to predict's \code{se.fit}, or NULL for
no value}

\item{...}{extra arguments (not used)}
}
\description{
Add fitted values, residuals, and other common outputs to
the value returned from \code{augment}.
}
\details{
In the case that a residuals or influence generic is not implemented for the
model, fail quietly.
}
