% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{augment.rma}
\alias{augment.rma}
\title{Augment data with values from a meta-analysis model}
\usage{
\method{augment}{rma}(x, ...)
}
\arguments{
\item{x}{An \code{rma} created by the \code{metafor} package.}

\item{...}{additional arguments}
}
\value{
a \code{tibble}
}
\description{
Augment the original data with model residuals, fitted values, and influence
statistics.
}
\examples{

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

augment(meta_analysis)

}
