% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-0-7-0.R
\name{bootstrap}
\alias{bootstrap}
\title{Set up bootstrap replicates of a dplyr operation}
\usage{
bootstrap(df, m, by_group = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{m}{number of bootstrap replicates to perform}

\item{by_group}{If \code{TRUE}, then bootstrap within each group if \code{df} is
a grouped tibble.}
}
\description{
The \code{bootstrap()} function is deprecated and will be removed from
an upcoming release of broom. For tidy resampling, please use the rsample
package instead. Functionality is no longer supported for this method.
}
\details{
This code originates from Hadley Wickham (with a few small
corrections) here:
\url{https://github.com/tidyverse/dplyr/issues/269}
}
\seealso{
Other deprecated: 
\code{\link{confint_tidy}()},
\code{\link{data.frame_tidiers}},
\code{\link{finish_glance}()},
\code{\link{fix_data_frame}()},
\code{\link{summary_tidiers}},
\code{\link{tidy.density}()},
\code{\link{tidy.dist}()},
\code{\link{tidy.ftable}()},
\code{\link{tidy.numeric}()}
}
\concept{deprecated}
