% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extras.R
\name{bootstrap}
\alias{bootstrap}
\title{Set up bootstrap replicates of a dplyr operation}
\usage{
bootstrap(df, m)
}
\arguments{
\item{df}{a data frame}

\item{m}{number of bootstrap replicates to perform}
}
\description{
Set up bootstrap replicates of a dplyr operation
}
\details{
This code originates from Hadley Wickham (with a few small
corrections) here:

https://github.com/hadley/dplyr/issues/269

Some examples can be found at

https://github.com/dgrtwo/broom/blob/master/vignettes/bootstrapping.Rmd
}
\examples{
library(dplyr)
mtcars \%>\% bootstrap(10) \%>\% do(tidy(lm(mpg ~ wt, .)))
}

