% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/broom.R
\docType{package}
\name{broom}
\alias{broom}
\alias{broom-package}
\title{Convert Statistical Analysis Objects into Tidy Data Frames}
\description{
Convert statistical analysis objects from R into tidy data frames,
so that they can more easily be combined, reshaped and otherwise processed
with tools like dplyr, tidyr and ggplot2. The package provides three S3
generics: tidy, which summarizes a model's statistical findings such as
coefficients of a regression; augment, which adds columns to the original
data such as predictions, residuals and cluster assignments; and glance,
which provides a one-row summary of model-level statistics.
}

