% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-package.R, R/broom.R
\docType{package}
\name{broom}
\alias{broom}
\alias{_PACKAGE}
\alias{broom-package}
\title{Convert Statistical Objects into Tidy Tibbles}
\description{
Convert statistical analysis objects from R into tidy tibbles,
so that they can more easily be combined, reshaped and otherwise processed
with tools like dplyr, tidyr and ggplot2. The package provides three S3
generics: tidy, which summarizes a model's statistical findings such as
coefficients of a regression; augment, which adds columns to the original
data such as predictions, residuals and cluster assignments; and glance,
which provides a one-row summary of model-level statistics.

Convert statistical analysis objects from R into tidy tibbles,
so that they can more easily be combined, reshaped and otherwise processed
with tools like dplyr, tidyr and ggplot2. The package provides three S3
generics: tidy, which summarizes a model's statistical findings such as
coefficients of a regression; augment, which adds columns to the original
data such as predictions, residuals and cluster assignments; and glance,
which provides a one-row summary of model-level statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://broom.tidyverse.org/}
  \item \url{http://github.com/tidymodels/broom}
  \item Report bugs at \url{http://github.com/tidymodels/broom/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

Authors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com}
}

Other contributors:
\itemize{
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) [contributor]
  \item Derek Chiu \email{dchiu@bccrc.ca} [contributor]
  \item Matthieu Gomez \email{mattg@princeton.edu} [contributor]
  \item Boris Demeshev \email{boris.demeshev@gmail.com} [contributor]
  \item Dieter Menne \email{dieter.menne@menne-biomed.de} [contributor]
  \item Benjamin Nutter \email{nutter@battelle.org} [contributor]
  \item Luke Johnston \email{luke.johnston@mail.utoronto.ca} [contributor]
  \item Ben Bolker \email{bolker@mcmaster.ca} [contributor]
  \item Francois Briatte \email{f.briatte@gmail.com} [contributor]
  \item Jeffrey Arnold \email{jeffrey.arnold@gmail.com} [contributor]
  \item Jonah Gabry \email{jsg2201@columbia.edu} [contributor]
  \item Luciano Selzer \email{luciano.selzer@gmail.com} [contributor]
  \item Gavin Simpson \email{ucfagls@gmail.com} [contributor]
  \item Jens Preussner \email{ jens.preussner@mpi-bn.mpg.de} [contributor]
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} [contributor]
  \item Hadley Wickham \email{hadley@rstudio.com} [contributor]
  \item Matthew Lincoln \email{matthew.d.lincoln@gmail.com} [contributor]
  \item Alessandro Gasparini \email{ag475@leicester.ac.uk} [contributor]
  \item Lukasz Komsta \email{lukasz.komsta@umlub.pl} [contributor]
  \item Frederick Novometsky [contributor]
  \item Wilson Freitas [contributor]
  \item Michelle Evans [contributor]
  \item Jason Cory Brunson \email{cornelioid@gmail.com} [contributor]
  \item Simon Jackson \email{drsimonjackson@gmail.com} [contributor]
  \item Ben Whalley \email{ben.whalley@plymouth.ac.uk} [contributor]
  \item Karissa Whiting \email{karissa.whiting@gmail.com} [contributor]
  \item Michael Kuehn \email{mkuehn10@gmail.com} [contributor]
  \item Jorge Cimentada \email{cimentadaj@gmail.com} [contributor]
  \item Erle Holgersen \email{erle.holgersen@gmail.com} [contributor]
  \item Karl Dunkle Werner [contributor]
  \item Ethan Christensen \email{christensen.ej@gmail.com} [contributor]
  \item Steven Pav \email{shabbychef@gmail.com} [contributor]
  \item Paul PJ \email{pjpaul.stephens@gmail.com} [contributor]
  \item Ben Schneider \email{benjamin.julius.schneider@gmail.com} [contributor]
  \item Patrick Kennedy \email{pkqstr@protonmail.com} [contributor]
  \item Lily Medina \email{lilymiru@gmail.com} [contributor]
  \item Brian Fannin \email{captain@pirategrunt.com} [contributor]
  \item Jason Muhlenkamp \email{jason.muhlenkamp@gmail.com} [contributor]
  \item Matt Lehman [contributor]
  \item Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID}) [contributor]
  \item Nic Crane [contributor]
  \item Andrew Bates [contributor]
  \item Vincent Arel-Bundock \email{vincent.arel-bundock@umontreal.ca} (\href{https://orcid.org/0000-0003-2042-7063}{ORCID}) [contributor]
  \item Hideaki Hayashi [contributor]
  \item Luis Tobalina [contributor]
  \item Annie Wang \email{anniewang.uc@gmail.com} [contributor]
  \item Wei Yang Tham \email{weiyang.tham@gmail.com} [contributor]
  \item Clara Wang \email{clara.wang.94@gmail.com} [contributor]
  \item Abby Smith \email{als1@u.northwestern.edu} (\href{https://orcid.org/0000-0002-3207-0375}{ORCID}) [contributor]
  \item Jasper Cooper \email{jaspercooper@gmail.com} (\href{https://orcid.org/0000-0002-8639-3188}{ORCID}) [contributor]
  \item E Auden Krauska \email{krauskae@gmail.com} (\href{https://orcid.org/0000-0002-1466-5850}{ORCID}) [contributor]
  \item Alex Wang \email{x249wang@uwaterloo.ca} [contributor]
  \item Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID}) [contributor]
  \item Charles Gray \email{charlestigray@gmail.com} (\href{https://orcid.org/0000-0002-9978-011X}{ORCID}) [contributor]
  \item Jared Wilber [contributor]
  \item Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID}) [contributor]
  \item Eduard Szoecs \email{eduardszoecs@gmail.com} [contributor]
  \item Frederik Aust \email{frederik.aust@uni-koeln.de} (\href{https://orcid.org/0000-0003-4900-788X}{ORCID}) [contributor]
  \item Angus Moore \email{angusmoore9@gmail.com} [contributor]
  \item Nick Williams \email{ntwilliams.personal@gmail.com} [contributor]
  \item Marius Barth \email{marius.barth.uni.koeln@gmail.com} (\href{https://orcid.org/0000-0002-3421-6665}{ORCID}) [contributor]
  \item Bruna Wundervald \email{brunadaviesw@gmail.com} (\href{https://orcid.org/0000-0001-8163-220X}{ORCID}) [contributor]
  \item Joyce Cahoon \email{joyceyu48@gmail.com} (\href{https://orcid.org/0000-0001-7217-4702}{ORCID}) [contributor]
  \item Kevin Zarca \email{kevin.zarca@gmail.com} [contributor]
  \item Shiro Kuriwaki \email{shirokuriwaki@gmail.com} (\href{https://orcid.org/0000-0002-5687-2647}{ORCID}) [contributor]
}

}
\keyword{internal}
