% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factanal_tidiers.R
\name{factanal_tidiers}
\alias{augment.factanal}
\alias{factanal_tidiers}
\alias{glance.factanal}
\alias{tidy.factanal}
\title{Tidying methods for a linear model}
\usage{
\method{tidy}{factanal}(x, ...)

\method{augment}{factanal}(x, data, ...)

\method{glance}{factanal}(x, ...)
}
\arguments{
\item{x}{\code{\link{factanal}} object}

\item{...}{Additional arguments, not used}

\item{data}{Original data}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy.factanal} returns one row for each variable used in the 
  analysis and the following columns:
  \item{variable}{The variable being estimated in the factor analysis}
  \item{uniqueness}{Proportion of residual, or unexplained variance}
  \item{flX}{Factor loading of term on factor X. There will be as many columns
  of this format as there were factors fitted.}

When \code{data} is not supplied \code{augment.factanal} returns one 
  row for each observation, with a factor score column added for each factor 
  X, (\code{.fsX}). This is because \code{\link{factanal}}, unlike other 
  stats methods like \code{\link{lm}}, does not retain the original data.

When \code{data} is supplied, \code{augment.factanal} returns one row for
each observation, with a factor score column added for each factor X,
(\code{.fsX}).

\code{glance.factanal} returns a one-row data.frame with the columns:
  \item{n.factors}{The number of fitted factors}
  \item{total.variance}{Total cumulative proportion of variance accounted for by all factors}
  \item{statistic}{Significance-test statistic}
  \item{p.value}{p-value from the significance test, describing whether the
  covariance matrix estimated from the factors is significantly different
  from the observed covariance matrix}
  \item{df}{Degrees of freedom used by the factor analysis}
  \item{n}{Sample size used in the analysis}
  \item{method}{The estimation method; always Maximum Likelihood, "mle"}
  \item{converged}{Whether the factor analysis converged}
}
\description{
These methods tidy the factor loadings of a factor analysis, conducted via 
\code{\link{factanal}}, into a summary, augment the original data with factor
scores, and construct a one-row glance of the model's statistics.
}
\examples{

mod <- factanal(mtcars, 3, scores = "regression")

glance(mod)
tidy(mod)
augment(mod)
augment(mod, mtcars)  # Must include original data if desired

# Factor loading plot

library(dplyr)
library(tidyr)
library(ggplot2)

tidy(mod) \%>\% 
    select(-uniqueness) \%>\% 
    gather(factor, loading, -variable) \%>\% 
    mutate(factor = gsub("fl", "", factor),
           sign = ifelse(sign(loading) == 1, "b", "a"),
           loading = abs(loading)) \%>\% 
    group_by(factor) \%>\% 
    arrange(factor, desc(loading)) \%>\% 
    mutate(order = 1:n()) \%>\% 
    ungroup() \%>\% 
    ggplot(aes(x = factor, y = loading, group = order)) +
    geom_bar(aes(fill = sign, alpha = loading),
             stat = "identity", position = "dodge", color = "black",
             show.legend = FALSE) +
    geom_text(aes(label = variable, group = order),
              position = position_dodge(width = 0.9),
              size = 3.5, angle = 45, hjust = -.4) +
    scale_fill_brewer(palette = "Set1") +
    theme_bw() +
    labs(
        x = "Factor",
        y = "Loading",
        title = "Positive (blue) and negative (red) factor loadings of all variables"
    )
}

