% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-0-7-0.R
\name{finish_glance}
\alias{finish_glance}
\title{(Deprecated) Add logLik, AIC, BIC, and other common measurements to a
glance of a prediction}
\usage{
finish_glance(ret, x)
}
\arguments{
\item{ret}{a one-row data frame (a partially complete glance)}

\item{x}{the prediction model}
}
\value{
a one-row data frame with additional columns added, such as
\item{logLik}{log likelihoods}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
This function is now deprecated in favor of using custom logic and
the appropriate \code{nobs()} method.
}
\seealso{
Other deprecated: 
\code{\link{bootstrap}()},
\code{\link{confint_tidy}()},
\code{\link{data.frame_tidiers}},
\code{\link{fix_data_frame}()},
\code{\link{summary_tidiers}},
\code{\link{tidy.density}()},
\code{\link{tidy.dist}()},
\code{\link{tidy.ftable}()},
\code{\link{tidy.numeric}()}
}
\concept{deprecated}
