% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-arima-tidiers.R
\name{glance.Arima}
\alias{glance.Arima}
\title{Glance at a(n) Arima object}
\usage{
\method{glance}{Arima}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Arima} created by \code{\link[stats:arima]{stats::arima()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.
}
\examples{

fit <- arima(lh, order = c(1, 0, 0))

tidy(fit)
glance(fit)

}
\seealso{
\code{\link[stats:arima]{stats::arima()}}

Other Arima tidiers: 
\code{\link{tidy.Arima}()}
}
\concept{Arima tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{AIC}{Akaike's Information Criterion for the model.}
  \item{BIC}{Bayesian Information Criterion for the model.}
  \item{logLik}{The log-likelihood of the model. [stats::logLik()] may be a useful reference.}
  \item{nobs}{Number of observations used.}
  \item{sigma}{Estimated standard error of the residuals.}

}
