% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg-tidiers.R
\name{glance.betareg}
\alias{glance.betareg}
\title{Glance at a(n) betareg object}
\usage{
\method{glance}{betareg}(x, ...)
}
\arguments{
\item{x}{A \code{betareg} object produced by a call to \code{\link[betareg:betareg]{betareg::betareg()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
\code{glance} returns a one-row tibble with columns:
\item{pseudo.r.squared}{the deviance of the null model}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{df.residual}{residual degrees of freedom}
\item{df.null}{degrees of freedom under the null}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[betareg:betareg]{betareg::betareg()}}
}
