% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm-tidiers.R
\name{glance.biglm}
\alias{glance.biglm}
\title{Glance at a(n) biglm object}
\usage{
\method{glance}{biglm}(x, ...)
}
\arguments{
\item{x}{A \code{biglm} object created by a call to \code{\link[biglm:biglm]{biglm::biglm()}} or
\code{\link[biglm:bigglm]{biglm::bigglm()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
\code{glance.biglm} returns a one-row data frame, with columns
\item{r.squared}{The percent of variance explained by the model}
\item{AIC}{the Akaike Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[biglm:biglm]{biglm::biglm()}}, \code{\link[biglm:bigglm]{biglm::bigglm()}}

Other biglm tidiers: \code{\link{tidy.biglm}}
}
\concept{biglm tidiers}
