% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg-tidiers.R
\name{glance.ivreg}
\alias{glance.ivreg}
\title{Glance at a(n) ivreg object}
\usage{
\method{glance}{ivreg}(x, diagnostics = FALSE, ...)
}
\arguments{
\item{x}{An \code{ivreg} object created by a call to \code{\link[ivreg:ivreg]{ivreg::ivreg()}}.}

\item{diagnostics}{Logical indicating whether to include diagnostics tests
from the first-stage regression; see Details. Defaults to \code{FALSE}.}

\item{...}{Further arguments to be passed to \code{\link[ivreg:ivreg.fit]{ivreg::ivreg.fit()}}, e.g.
specifying an alternate variance-covariance structure. See Examples.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modeling
function. This includes the name of the modeling function or any
arguments passed to the modeling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.

Glance returns the same number of columns regardless of whether the
model matrix is rank-deficient or not. If so, entries in columns
that no longer have a well-defined value are filled in with an \code{NA}
of the appropriate type.
}
\details{
As of 0.7.1, \code{glance.ivreg(..., diagnostics = TRUE)} returns three
diagnostics tests from the first-stage regression. These are 1) a weak
instruments F-test, a 2) Wu-Hausman chi-squared test of endogeneity, and 3)
Sargan's J-test of overidentifying restrictions. Note that the latter test
requires there to be more instruments than endogenous regressors. The Sargan
test values will automatically return \code{NA} if this condition is not met.
}
\examples{

library(ivreg)

data("CigaretteDemand", package = "ivreg")

m <- ivreg(log(packs) ~ log(rprice) + log(rincome) | salestax + log(rincome),
           data = CigaretteDemand)

summary(m)

tidy(m)
tidy(m, conf.int = TRUE)

# Specify alternate error structures by passing through an appropriate
# "vcov" argument. Here we use HC (i.e. robust) standard errors from the
# sandwich package.
tidy(m, vcov = sandwich::vcovHC)

# By default, the stage-2 component of the 2SLS regression is returned and
# tidied. But we can easily obtain the stage-1 component instead.
tidy(m, component = "stage1")

# If only the instrumental variables from stage-1 are desired, e.g. as
# supplementary information for a regression table.
tidy(m, component = "instruments")

augment(m)
augment(m, se_fit = TRUE)
augment(m, se_fit = TRUE, interval = "confidence")
augment(m, newdata = CigaretteDemand[1:10, ], interval = "prediction")

# As with tidy.ivreg, we can specify alternate error structures by passing
# on a "vcov" argument from the sandwich package that ivreg objects 
# understand.
augment(m, se_fit = TRUE, vcov = sandwich::vcovHC)

# The "interval" and "vcov" arguments above can be combined. Among other
# things, this can be useful for visually inspecting the impact of different 
# standard errors on our model predictions.
library(ggplot2)
ggplot(
  data = augment(m, interval = 'confidence'), 
  aes(x=.rownames, y=.fitted, ymin=.lower, ymax=.upper)
  ) + 
  geom_pointrange() +
  ylim(4, 5) +
  labs(title = "Predicted values; regular standard errors")
ggplot(
  data = augment(m, interval = 'confidence', vcov = sandwich::vcovHC),
  aes(x=.rownames, y=.fitted, ymin=.lower, ymax=.upper)
  ) + 
  geom_pointrange() +
  ylim(4, 5) +
  labs(title = "Predicted values; robust standard errors")

glance(m)
# Include stage-1 diagnostics tests
glance(m, diagnostics = TRUE)
# Specify alternate error structure (e.g. HC robust standard errors)
glance(m, vcov = sandwich::vcovHC)
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[AER:ivreg]{AER::ivreg()}}

Other ivreg tidiers: 
\code{\link{augment.ivreg}()},
\code{\link{tidy.ivreg}()}
}
\concept{ivreg tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with exactly one row and columns:
  \item{adj.r.squared}{Adjusted R squared statistic, which is like the R squared statistic except taking degrees of freedom into account.}
  \item{df}{Degrees of freedom used by the model.}
  \item{df.residual}{Residual degrees of freedom.}
  \item{nobs}{Number of observations used.}
  \item{r.squared}{R squared statistic, or the percent of variation explained by the model. Also known as the coefficient of determination.}
  \item{sigma}{Estimated standard error of the residuals.}
  \item{statistic}{Wald test statistic.}
  \item{p.value}{P-value for the Wald test.}
  \item{statistic.weak.instr}{Weak instrument(s) F-test.}
  \item{p.value.weak.instr}{P-value for weak instrument(s) test.}
  \item{statistic.Sargan}{Sargan J-statistic of overidentifying restrictions.}
  \item{p.value.Sargan}{P-value for Sargan J-test.}
  \item{statistic.Wu.Hausman}{Wu-Hausman chi-squared test of endogeneity.}
  \item{p.value.Wu.Hausman}{P-value for Wu-Hausman test.}

}
