% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-kmeans-tidiers.R
\name{glance.kmeans}
\alias{glance.kmeans}
\title{Glance at a(n) kmeans object}
\usage{
\method{glance}{kmeans}(x, ...)
}
\arguments{
\item{x}{A \code{kmeans} object created by \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}

\item{data}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{totss}{The total sum of squares}
\item{tot.withinss}{The total within-cluster sum of squares}
\item{betweenss}{The total between-cluster sum of squares}
\item{iter}{The numbr of (outer) iterations}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[stats:kmeans]{stats::kmeans()}}

Other kmeans tidiers: \code{\link{augment.kmeans}},
  \code{\link{tidy.kmeans}}
}
\concept{kmeans tidiers}
