% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-nls-tidiers.R
\name{glance.nls}
\alias{glance.nls}
\title{Glance at a(n) nls object}
\usage{
\method{glance}{nls}(x, ...)
}
\arguments{
\item{x}{An \code{nls} object returned from \code{\link[stats:nls]{stats::nls()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{sigma}{the square root of the estimated residual variance}
\item{isConv}{whether the fit successfully converged}
\item{finTol}{the achieved convergence tolerance}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\link{tidy}, \code{\link[stats:nls]{stats::nls()}}

Other nls tidiers: \code{\link{augment.nls}},
  \code{\link{tidy.nls}}
}
\concept{nls tidiers}
