% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance}
\alias{glance}
\title{Construct a single row summary "glance" of a model, fit, or other
object}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\description{
glance methods always return either a one-row data frame (except on NULL, which
returns an empty data frame)
}
