% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{glance.rma}
\alias{glance.rma}
\title{Glance at a(n) rma object}
\usage{
\method{glance}{rma}(x, ...)
}
\arguments{
\item{x}{An \code{rma} created by the \code{metafor} package.}

\item{...}{Additional arguments}
}
\value{
a \code{tibble}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.
}
\examples{

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

glance(meta_analysis)

}
