% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{glance.rma}
\alias{glance.rma}
\title{Construct a one-row summary of meta-analysis model fit statistics.}
\usage{
\method{glance}{rma}(x, ...)
}
\arguments{
\item{x}{a meta-analysis object. Currently supports \code{rma.uni} from the
\code{metafor} package.}

\item{...}{additional arguments}
}
\value{
a \code{data.frame}
}
\description{
\code{glance()} computes a one-row summary of  meta-analysis objects,
including estimates of heterogenity and model fit.
}
\examples{

example_ma \%>\%
  glance()

}
