% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustbase-glmrob-tidiers.R
\name{glance.glmrob}
\alias{glance.glmrob}
\title{Glance at a(n) glmrob object}
\usage{
\method{glance}{glmrob}(x, ...)
}
\arguments{
\item{x}{Unused.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}

\item{...}{Unused.}
}
\description{
Glance accepts a model object and returns a \code{\link[tibble:tibble]{tibble::tibble()}}
with exactly one row of model summaries. The summaries are typically
goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modelling
function. This includes the name of the modelling function or any
arguments passed to the modelling function.

Glance does not calculate summary measures. Rather, it farms out these
computations to appropriate methods and gathers the results together.
Sometimes a goodness of fit measure will be undefined. In these cases
the measure will be reported as \code{NA}.

\code{glance.glmrob()} has not yet been implemented in broom. The
\pkg{robustbase} package does not provide the functionality necessary to
implement a glance method.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}. For tidiers for robust models from the \pkg{robust} package
see \code{\link[=tidy.lmRob]{tidy.lmRob()}}.
}
\seealso{
\code{\link[robustbase:glmrob]{robustbase::glmrob()}}

Other robustbase tidiers: \code{\link{augment.glmrob}},
  \code{\link{augment.lmrob}}, \code{\link{glance.lmrob}},
  \code{\link{tidy.glmrob}}, \code{\link{tidy.lmrob}}
}
\concept{robustbase tidiers}
