% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedglm-tidiers.R
\name{glance.speedlm}
\alias{glance.speedlm}
\title{Glance at a(n) speedlm object}
\usage{
\method{glance}{speedlm}(x, ...)
}
\arguments{
\item{x}{A \code{speedlm} object returned from \code{\link[speedglm:speedlm]{speedglm::speedlm()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{r.squared}{The percent of variance explained by the model}
\item{adj.r.squared}{r.squared adjusted based on the degrees of freedom}
\item{statistic}{F-statistic}
\item{p.value}{p-value from the F test, describing whether the full
regression is significant}
\item{df}{Degrees of freedom used by the coefficients}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Description of what a \code{glance} method does goes here.
}
\seealso{
\code{\link[speedglm:speedlm]{speedglm::speedlm()}}

Other speedlm tidiers: \code{\link{augment.speedlm}},
  \code{\link{tidy.speedlm}}
}
\concept{speedlm tidiers}
