% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glm-tidiers}
\alias{glance.glm}
\alias{glm-tidiers}
\title{Tidying methods for a glm object}
\usage{
\method{glance}{glm}(x, ...)
}
\arguments{
\item{x}{glm object}

\item{...}{extra arguments, not used}
}
\value{
\code{tidy} and \code{augment} return the same values as do
\code{\link{tidy.lm}} and \code{\link{augment.lm}}.

\code{glance} returns a one-row data.frame with the columns
  \item{deviance}{Minus twice the maximized log-likelihood}
  \item{null.deviance}{The deviance of the null model}
  \item{df.residual}{The residual degrees of freedom}
  \item{df.null}{The residual degrees of freedom for the null model}
  \item{logLik}{The data's log-likelihood under the model}
  \item{AIC}{The Akaike Information Criterion}
  \item{BIC}{The Bayesian Information Criterion}
}
\description{
Tidy a \code{glm} object. The \code{tidy} and \code{augment} methods are handled
by \code{\link{lm-tidiers}}.
}
\seealso{
\code{\link{tidy.lm}} and \code{\link{augment.lm}}. Also \code{\link{glm}}, which
computes the values reported by the \code{glance} method.
}

