% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glm_tidiers}
\alias{augment.glm}
\alias{glance.glm}
\alias{glm_tidiers}
\title{Tidying methods for a glm object}
\usage{
\method{glance}{glm}(x, ...)

\method{augment}{glm}(x, data = x$model, newdata = NULL, ...)
}
\arguments{
\item{x}{glm object}

\item{...}{extra arguments, not used}
}
\value{
\code{tidy} and \code{augment} return the same values as do
\code{\link{tidy.lm}} and \code{\link{augment.lm}}.

\code{glance} returns a one-row data.frame with the columns
  \item{null.deviance}{the deviance of the null model}
  \item{df.null}{the residual degrees of freedom for the null model}
  \item{logLik}{the data's log-likelihood under the model}
  \item{AIC}{the Akaike Information Criterion}
  \item{BIC}{the Bayesian Information Criterion}
  \item{deviance}{deviance}
  \item{df.residual}{residual degrees of freedom}

When \code{newdata} is not supplied \code{augment.glm} returns one row for each observation,
with nine columns added to the original data:
  \item{.hat}{Diagonal of the hat matrix}
  \item{.sigma}{Estimate of residual standard deviation when
    corresponding observation is dropped from model}
  \item{.cooksd}{Cooks distance, \code{\link{cooks.distance}}}
  \item{.fitted.link}{Fitted values of model}
  \item{.se.fit.link}{Standard errors of fitted values}
  \item{.fitted.response}{Fitted values of model}
  \item{.se.fit.response}{Standard errors of fitted values}
  \item{.resid}{Residuals}
  \item{.stdresid}{Standardised residuals}
When \code{newdata} is supplied  \code{augment.glm} returns one row for each observation,
with two columns added to the new data:
  \item{.fitted.link}{Fitted values of model}
  \item{.se.fit.link}{Standard errors of fitted values}
  \item{.fitted.response}{Fitted values of model}
  \item{.se.fit.response}{Standard errors of fitted values}
}
\description{
Tidy a \code{glm} object. The \code{tidy} and \code{augment} methods are handled
by \code{\link{lm_tidiers}}.
}
\details{
Code and documentation for \code{augment.glm} originated in the
ggplot2 package, where it was called \code{fortify.lm}
}
\examples{
g <- glm(am ~ mpg, mtcars, family = "binomial")
glance(g)
}
\seealso{
\code{\link{tidy.lm}} and \code{\link{augment.lm}}. Also \code{\link{glm}}, which
computes the values reported by the \code{glance} method.
}

