% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_tidiers.R
\name{lavaan_tidiers}
\alias{lavaan_tidiers}
\alias{tidy.lavaan}
\alias{glance.lavaan}
\title{Tidying methods for lavaan models}
\usage{
\method{tidy}{lavaan}(x, conf.int = TRUE, conf.level = 0.95, ...)

\method{glance}{lavaan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan}, such as those from \code{\link[lavaan:cfa]{lavaan::cfa()}},
or \code{\link[lavaan:sem]{lavaan::sem()}}}

\item{conf.int}{Logical indicating if a confidence intervals are desired.}

\item{conf.level}{Confidence level to use. Default is 0.95.}

\item{...}{For \code{tidy}, additional arguments passed to
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. Ignored for \code{glance}.}
}
\value{
The \code{tidy} method returns a tibble with one row for each
estimated parameter and columns:

\item{term}{The result of paste(lhs, op, rhs)}
\item{op}{The operator in the model syntax (e.g. \code{~~} for covariances, or
\code{~} for regression parameters)}
\item{group}{The group (if specified) in the lavaan model}
\item{estimate}{The parameter estimate (may be standardized)}
\item{std.error}{}
\item{statistic}{The z value returned by \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}}
\item{p.value}{}
\item{conf.low}{}
\item{conf.high}{}
\item{std.lv}{Standardized estimates based on the variances of the
(continuous) latent variables only}
\item{std.all}{Standardized estimates based on both the variances
of both (continuous) observed and latent variables.}
\item{std.nox}{Standardized estimates based on both the variances
of both (continuous) observed and latent variables, but not the
variances of exogenous covariates.}

The \code{glance} method returns a single row dataset giving model fit
and related information, including:

\item{chisq}{Model chi squared}
\item{npar}{Number of parameters in the model}
\item{rmsea}{Root mean square error of approximation}
\item{rmsea.conf.high}{95 percent upper bound on RMSEA}
\item{srmr}{Standardised root mean residual}
\item{agfi}{Adjusted goodness of fit}
\item{cfi}{Comparative fit index}
\item{tli}{Tucker Lewis index}
\item{aic}{Akaike information criterion}
\item{bic}{Bayesian information criterion}
\item{ngroups}{Number of groups in model}
\item{nobs}{Number of observations included}
\item{norig}{Number of observation in the original dataset}
\item{nexcluded}{Number of excluded observations}
\item{converged}{Logical - Did the model converge}
\item{estimator}{Estimator used}
\item{missing_method}{Method for eliminating missing data}

For further recommendations on reporting SEM and CFA models see Schreiber, J. B. (2017). Update to core reporting practices in structural equation modeling. Research in Social and Administrative Pharmacy, 13(3), 634-643. https://doi.org/10.1016/j.sapharm.2016.06.006
}
\description{
These methods tidy the coefficients of lavaan CFA and SEM models.
}
\examples{

if (require("lavaan")) {
 cfa.fit <- cfa('F =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9',
                data = HolzingerSwineford1939, group = "school")
 tidy(cfa.fit)
}

if (require("lavaan", quietly = TRUE)) {

 cfa.fit <- cfa(
   'F =~ x1 + x2 + x3 + x4 + x5',
   data = HolzingerSwineford1939, group = "school"
 )
 glance(cfa.fit)

}
}
