% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-tidiers.R
\name{mcmc_tidiers}
\alias{mcmc_tidiers}
\alias{tidyMCMC}
\alias{tidy.rjags}
\alias{tidy.stanfit}
\title{Tidying methods for MCMC (Stan, JAGS, etc.) fits}
\usage{
tidyMCMC(x, pars, estimate.method = "mean", conf.int = FALSE,
  conf.level = 0.95, conf.method = "quantile", droppars = "lp__",
  rhat = FALSE, ess = FALSE, ...)

\method{tidy}{rjags}(x, pars, estimate.method = "mean",
  conf.int = FALSE, conf.level = 0.95, conf.method = "quantile", ...)

\method{tidy}{stanfit}(...)
}
\arguments{
\item{x}{an object of class \sQuote{"stanfit"}}

\item{pars}{(character) specification of which parameters to include}

\item{estimate.method}{method for computing point estimate ("mean" or median")}

\item{conf.int}{(logical) include confidence interval?}

\item{conf.level}{probability level for CI}

\item{conf.method}{method for computing confidence intervals
("quantile" or "HPDinterval")}

\item{droppars}{Parameters not to include in the output (such
as log-probability information)}

\item{rhat, ess}{(logical) include Rhat and/or effective sample size estimates?}

\item{...}{unused}
}
\description{
MCMC tidiers are deprecated.
}
