% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjoint_tidiers.R
\name{mjoint_tidiers}
\alias{mjoint_tidiers}
\alias{tidy.mjoint}
\alias{glance.mjoint}
\title{Tidying methods for joint models for time-to-event data and multivariate longitudinal data}
\usage{
\method{tidy}{mjoint}(x, component = "survival", bootSE = NULL, ...)

\method{glance}{mjoint}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mjoint}.}

\item{component}{Either \code{survival} (the survival component of the model, default) or \code{longitudinal} (the longitudinal component).}

\item{bootSE}{An object of class \code{bootSE} for the corresponding model. If \code{bootSE = NULL} (the default), the function will attempt to utilize approximate standard error estimates (if available) calculated from the empirical information matrix.}

\item{...}{extra arguments (not used)}
}
\value{
All tidying methods return a \code{data.frame} without rownames.
The structure depends on the method chosen.

\code{tidy} returns one row for each estimated fixed effect depending on the \code{component} parameter. It contains the following  columns:
  \item{term}{The term being estimated}
  \item{estimate}{Estimated value}
  \item{std.error}{Standard error}
  \item{statistic}{Z-statistic}
  \item{p.value}{P-value computed from Z-statistic}

\code{glance} returns one row with the columns
  \item{sigma2_j}{the square root of the estimated residual variance for the j-th longitudinal process}
  \item{AIC}{the Akaike Information Criterion}
  \item{BIC}{the Bayesian Information Criterion}
  \item{logLik}{the data's log-likelihood under the model}
}
\description{
These methods tidy the coefficients of joint models for time-to-event data and multivariate
longitudinal data of the \code{mjoint} class from the \code{joineRML} package.
}
\examples{
\dontrun{
# Fit a joint model with bivariate longitudinal outcomes
library(joineRML)
data(heart.valve)
hvd <- heart.valve[!is.na(heart.valve$log.grad) & !is.na(heart.valve$log.lvmi) & heart.valve$Num <= 50, ]
fit <- mjoint(
  formLongFixed = list("grad" = log.grad ~ time + sex + hs,
                       "lvmi" = log.lvmi ~ time + sex),
  formLongRandom = list("grad" = ~ 1 | num,
                        "lvmi" = ~ time | num),
  formSurv = Surv(fuyrs, status) ~ age,
  data = hvd,
  inits = list("gamma" = c(0.11, 1.51, 0.80)),
  timeVar = "time")

# Extract the survival fixed effects
tidy(fit)

# Extract the longitudinal fixed effects
tidy(fit, component = "longitudinal")

# Extract model statistics
glimpse(fit)
}

}
