% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-package.R, R/broom.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\seealso{
\code{\link[=augment.lm]{augment.lm()}}

\code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[=glance.lm]{glance.lm()}}

\code{\link[=augment.lm]{augment.lm()}}

\code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[=glance.lm]{glance.lm()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}, \code{\link[generics]{tidy}}}
}}

