% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-tidiers.R
\name{sp_tidiers}
\alias{sp_tidiers}
\alias{tidy.SpatialPolygonsDataFrame}
\alias{tidy.SpatialPolygons}
\alias{tidy.Polygons}
\alias{tidy.Polygon}
\alias{tidy.SpatialLinesDataFrame}
\alias{tidy.Lines}
\alias{tidy.Line}
\title{Tidy a(n) SpatialPolygonsDataFrame object}
\usage{
\method{tidy}{SpatialPolygonsDataFrame}(x, region = NULL, ...)

\method{tidy}{SpatialPolygons}(x, ...)

\method{tidy}{Polygons}(x, ...)

\method{tidy}{Polygon}(x, ...)

\method{tidy}{SpatialLinesDataFrame}(x, ...)

\method{tidy}{Lines}(x, ...)

\method{tidy}{Line}(x, ...)
}
\arguments{
\item{x}{A \code{SpatialPolygonsDataFrame}, \code{SpatialPolygons}, \code{Polygons},
\code{Polygon}, \code{SpatialLinesDataFrame}, \code{Lines} or \code{Line} object.}

\item{region}{name of variable used to split up regions}

\item{...}{not used by this method}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.

Note that the \code{sf} package now defines tidy spatial objects
and is the recommend approach to spatial data. \code{sp} tidiers are likely
to be deprecated in the near future in favor of \code{sf::st_as_sf()}.
Development of \code{sp} tidiers has halted in \code{broom}.
}
