% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-summary-tidiers.R
\name{summary_tidiers}
\alias{summary_tidiers}
\alias{tidy.summaryDefault}
\alias{glance.summaryDefault}
\title{Tidy/glance a(n) summaryDefault object}
\usage{
\method{tidy}{summaryDefault}(x, ...)

\method{glance}{summaryDefault}(x, ...)
}
\arguments{
\item{x}{A \code{summaryDefault} object, created by calling \code{\link[=summary]{summary()}} on a
vector.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:
\item{minimum}{Minimum value in original vector.}
\item{q1}{First quartile of original vector.}
\item{median}{Median of original vector.}
\item{mean}{Mean of original vector.}
\item{q3}{Third quartile of original vector.}
\item{maximum}{Maximum value in original vector.}
\item{na}{Number of \code{NA} values in original vector. Column present only
when original vector had at least one \code{NA} entry.}
}
\description{
For models that have only a single component, the \code{\link[=tidy]{tidy()}} and
\code{\link[=glance]{glance()}} methods are identical. Please see the documentation for both
of those methods.
}
\examples{

v <- rnorm(1000)
s <- summary(v)
s

tidy(s)
glance(s)

v2 <- c(v,NA)
tidy(summary(v2))

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=summary]{summary()}}
}
