% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-irlba.R, R/list-svd-tidiers.R
\name{svd_tidiers}
\alias{svd_tidiers}
\alias{tidy_irlba}
\alias{svd_tidiers}
\alias{tidy_svd}
\title{Tidy an irlba list}
\usage{
tidy_irlba(x, ...)

tidy_svd(x, matrix = "u", ...)
}
\arguments{
\item{x}{list containing d, u, v components, returned from \code{svd}}

\item{...}{Extra arguments (not used)}

\item{matrix}{which of the u, d or v matrix to tidy}
}
\value{
An SVD object contains a decomposition into u, d, and v matrices,
such that\preformatted{u %*% diag(d) %*% t(v)
}

gives the original matrix. This tidier gives a choice of which matrix
to tidy.

When \code{matrix = "u"}, each observation represents one pair of row and
principal component, with variables:
\item{row}{Number of the row in the original data being described}
\item{PC}{Principal component}
\item{value}{Loading of this principal component for this row}

When \code{matrix = "d"}, each observation represents one principal component,
with variables:
\item{\code{PC}}{An integer vector indicating the principal component}
\item{\code{std.dev}}{Standard deviation explained by this PC}
\item{\code{percent}}{Percentage of variation explained. This will be
inaccurate for irlba objects due to the SVD truncation.}
\item{\code{cumulative}}{Cumulative percentage of variation explained}

When \code{matrix = "v"}, each observation represents a pair of a principal
component and a column of the original matrix, with variables:
\item{column}{Column of original matrix described}
\item{PC}{Principal component}
\item{value}{Value of this PC for this column}
}
\description{
A very thin wrapper around \code{\link[=tidy_svd]{tidy_svd()}}.

These methods tidy the U, D, and V matrices returned by the
\code{\link[=svd]{svd()}} functions into a tidy format. Because \code{svd} returns
lists without classes, this function has to be called by
\code{\link[=tidy.list]{tidy.list()}} when it recognizes a list as an SVD.
}
\details{
Note that SVD is only equivalent to PCA on centered data.
}
\examples{

mat <- scale(as.matrix(iris[, 1:4]))
s <- svd(mat)

tidy_u <- tidy(s, matrix = "u")
head(tidy_u)

tidy_d <- tidy(s, matrix = "d")
tidy_d

tidy_v <- tidy(s, matrix = "v")
head(tidy_v)

library(ggplot2)
library(dplyr)

ggplot(tidy_d, aes(PC, percent)) +
    geom_point() +
    ylab("\% of variance explained")

tidy_u \%>\%
    mutate(Species = iris$Species[row]) \%>\%
    ggplot(aes(Species, value)) +
    geom_boxplot() +
    facet_wrap(~ PC, scale = "free_y")
}
\seealso{
\code{\link[=svd]{svd()}}, \code{\link[irlba:irlba]{irlba::irlba()}}, \code{\link[=tidy.list]{tidy.list()}}, \code{\link[=tidy_irlba]{tidy_irlba()}}
}
