% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-tidiers.R
\name{tidy.rma}
\alias{tidy.rma}
\title{Tidying methods for meta-analyis objects}
\usage{
\method{tidy}{rma}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, include_studies = TRUE, measure = "GEN", ...)
}
\arguments{
\item{x}{An \code{rma} created by the \code{metafor} package.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{include_studies}{Logical. Should individual studies be included in the
output?}

\item{measure}{Measure type. See \code{\link[metafor:rma]{metafor::rma()}}}

\item{...}{Additional arguments}
}
\value{
A \code{tibble}
}
\description{
These methods tidy the results of meta-analysis objects from the metafor package
}
\examples{

library(metafor)

df <-
  escalc(
    measure = "RR",
    ai = tpos,
    bi = tneg,
    ci = cpos,
    di = cneg,
    data = dat.bcg
  )

meta_analysis <- rma(yi, vi, data = df, method = "EB")

tidy(meta_analysis)

}
