% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-anova.R
\name{tidy.aov}
\alias{tidy.aov}
\title{Tidy a(n) aov object}
\usage{
\method{tidy}{aov}(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{An \code{aov} object, such as those created by \code{\link[stats:aov]{stats::aov()}}.}

\item{intercept}{A logical indicating whether information on the intercept
ought to be included. Passed to \code{\link[stats:summary.aov]{stats::summary.aov()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Two exceptions here are:
\itemize{
\item \code{tidy()} methods will warn when supplied an \code{exponentiate} argument if
it will be ignored.
\item \code{augment()} methods will warn when supplied a \code{newdata} argument if it
will be ignored.
}}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies across models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
The \code{term} column of an ANOVA table can come with leading or
trailing whitespace, which this tidying method trims.

For documentation on the tidier for \code{\link[car:leveneTest]{car::leveneTest()}} output, see
\code{\link[=tidy.leveneTest]{tidy.leveneTest()}}
}
\examples{

a <- aov(mpg ~ wt + qsec + disp, mtcars)
tidy(a)
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[stats:aov]{stats::aov()}}

Other anova tidiers: 
\code{\link{glance.anova}()},
\code{\link{glance.aov}()},
\code{\link{tidy.TukeyHSD}()},
\code{\link{tidy.anova}()},
\code{\link{tidy.aovlist}()},
\code{\link{tidy.manova}()}
}
\concept{anova tidiers}
