% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.density}
\alias{tidy.density}
\title{tidy a density objet}
\usage{
\method{tidy}{density}(x, ...)
}
\arguments{
\item{x}{an object of class "density"}

\item{...}{extra arguments (not used)}
}
\value{
a data frame with "x" and "y" columns
}
\description{
Given a "density" object, returns a tidy data frame with two
columns: points x where the density is estimated, points y
for the estimate
}

