% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-0-7-0.R
\name{tidy.density}
\alias{tidy.density}
\title{(Deprecated) Tidy density objects}
\usage{
\method{tidy}{density}(x, ...)
}
\arguments{
\item{x}{A \code{density} object returned from \code{\link[stats:density]{stats::density()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with two columns: points \code{x} where the density
is estimated, and estimated density \code{y}.
}
\description{
(Deprecated) Tidy density objects
}
\seealso{
Other deprecated: 
\code{\link{bootstrap}()},
\code{\link{confint_tidy}()},
\code{\link{data.frame_tidiers}},
\code{\link{finish_glance}()},
\code{\link{fix_data_frame}()},
\code{\link{summary_tidiers}},
\code{\link{tidy.dist}()},
\code{\link{tidy.ftable}()},
\code{\link{tidy.numeric}()}
}
\concept{deprecated}
