% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-0-7-0.R
\name{tidy.ftable}
\alias{tidy.ftable}
\title{(Deprecated) Tidy ftable objects}
\usage{
\method{tidy}{ftable}(x, ...)
}
\arguments{
\item{x}{An \code{ftable} object returned from \code{\link[stats:ftable]{stats::ftable()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Two exceptions here are:
\itemize{
\item \code{tidy()} methods will warn when supplied an \code{exponentiate} argument if
it will be ignored.
\item \code{augment()} methods will warn when supplied a \code{newdata} argument if it
will be ignored.
}}
}
\value{
An ftable contains a "flat" contingency table. This melts it into a
\link[tibble:tibble]{tibble::tibble} with one column for each variable, then a \code{Freq}
column.
}
\description{
This function is deprecated. Please use \code{\link[tibble:as_tibble]{tibble::as_tibble()}} instead.
}
\seealso{
Other deprecated: 
\code{\link{bootstrap}()},
\code{\link{confint_tidy}()},
\code{\link{data.frame_tidiers}},
\code{\link{finish_glance}()},
\code{\link{fix_data_frame}()},
\code{\link{summary_tidiers}},
\code{\link{tidy.density}()},
\code{\link{tidy.dist}()},
\code{\link{tidy.numeric}()}
}
\concept{deprecated}
