% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats_tidiers.R
\name{tidy.ftable}
\alias{tidy.ftable}
\title{tidy an ftable object}
\usage{
\method{tidy}{ftable}(x, ...)
}
\arguments{
\item{x}{An object of class "ftable"}

\item{...}{Extra arguments (not used)}
}
\description{
An ftable contains a "flat" contingency table. This melts it into a
data.frame with one column for each variable, then a \code{Freq}
column. It directly uses the \code{stats:::as.data.frame.ftable} function
}
\examples{
tidy(ftable(Titanic, row.vars = 1:3))
}
\seealso{
\code{\link{ftable}}
}

