% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-tidiers.R
\name{tidy.gam}
\alias{tidy.gam}
\alias{mgcv_tidiers}
\alias{gam_tidiers}
\title{Tidy a(n) gam object}
\usage{
\method{tidy}{gam}(x, parametric = FALSE, ...)
}
\arguments{
\item{x}{A \code{gam} object returned from a call to \code{\link[mgcv:gam]{mgcv::gam()}}.}

\item{parametric}{Logical indicating if parametric or smooth terms should
be tidied. Defaults to \code{FALSE}, meaning that smooth terms are tidied
by default.}
}
\description{
Description of what a \code{tidy} method does goes here.
}
\details{
To tidy \code{Gam} objects created by calls to \code{\link[gam:gam]{gam::gam()}},
see \code{\link[=tidy.Gam]{tidy.Gam()}}.
}
\examples{

if (requireNamespace("mgcv", quietly = TRUE)) {
  g <- mgcv::gam(mpg ~ s(hp) + am + qsec, data = mtcars)
  
  tidy(g)
  tidy(g, parametric = TRUE)
  glance(g)
}

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[mgcv:gam]{mgcv::gam()}}, \code{\link[=tidy.Gam]{tidy.Gam()}}

Other mgcv tidiers: \code{\link{glance.gam}}
}
\concept{mgcv tidiers}
