% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg-tidiers.R
\name{tidy.ivreg}
\alias{tidy.ivreg}
\alias{ivreg_tidiers}
\title{Tidy a(n) ivreg object}
\usage{
\method{tidy}{ivreg}(
  x,
  conf.int = FALSE,
  conf.level = 0.95,
  component = c("stage2", "stage1", "instruments"),
  ...
)
}
\arguments{
\item{x}{An \code{ivreg} object created by a call to \code{\link[ivreg:ivreg]{ivreg::ivreg()}}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{component}{Character indicating which component of the 2SLS regression
to tidy. One of "stage2" (the default), "stage1", or "instruments". See
examples.}

\item{...}{Further arguments to be passed to \code{\link[ivreg:ivreg.fit]{ivreg::ivreg.fit()}}, e.g.
specifying an alternate variance-covariance structure. Currently, only
supported for the default "stage2" component (above). See examples.}
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies across models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\examples{

library(ivreg)

data("CigaretteDemand", package = "ivreg")

m <- ivreg(log(packs) ~ log(rprice) + log(rincome) | salestax + log(rincome),
           data = CigaretteDemand)

summary(m)

tidy(m)
tidy(m, conf.int = TRUE)

# Specify alternate error structures by passing through an appropriate
# "vcov" argument. Here we use HC (i.e. robust) standard errors from the
# sandwich package.
tidy(m, vcov = sandwich::vcovHC)

# By default, the stage-2 component of the 2SLS regression is returned and
# tidied. But we can easily obtain the stage-1 component instead.
tidy(m, component = "stage1")

# If only the instrumental variables from stage-1 are desired, e.g. as
# supplementary information for a regression table.
tidy(m, component = "instruments")

augment(m)
augment(m, se_fit = TRUE)
augment(m, se_fit = TRUE, interval = "confidence")
augment(m, newdata = CigaretteDemand[1:10, ], interval = "prediction")

# As with tidy.ivreg, we can specify alternate error structures by passing
# on a "vcov" argument from the sandwich package that ivreg objects 
# understand.
augment(m, se_fit = TRUE, vcov = sandwich::vcovHC)

# The "interval" and "vcov" arguments above can be combined. Among other
# things, this can be useful for visually inspecting the impact of different 
# standard errors on our model predictions.
library(ggplot2)
ggplot(
  data = augment(m, interval = 'confidence'), 
  aes(x=.rownames, y=.fitted, ymin=.lower, ymax=.upper)
  ) + 
  geom_pointrange() +
  ylim(4, 5) +
  labs(title = "Predicted values; regular standard errors")
ggplot(
  data = augment(m, interval = 'confidence', vcov = sandwich::vcovHC),
  aes(x=.rownames, y=.fitted, ymin=.lower, ymax=.upper)
  ) + 
  geom_pointrange() +
  ylim(4, 5) +
  labs(title = "Predicted values; robust standard errors")

glance(m)
# Include stage-1 diagnostics tests
glance(m, diagnostics = TRUE)
# Specify alternate error structure (e.g. HC robust standard errors)
glance(m, vcov = sandwich::vcovHC)
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[ivreg:ivreg]{ivreg::ivreg()}}

Other ivreg tidiers: 
\code{\link{augment.ivreg}()},
\code{\link{glance.ivreg}()}
}
\concept{ivreg tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{conf.high}{Upper bound on the confidence interval for the estimate.}
  \item{conf.low}{Lower bound on the confidence interval for the estimate.}
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}

}
