% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tidy.lm}
\alias{tidy.lm}
\title{tidy a linear model by returning a data frame version of the coefficients
table}
\usage{
\method{tidy}{lm}(x, ...)
}
\arguments{
\item{x}{An object of class "lm"}

\item{...}{extra arguments (not used)}
}
\value{
a data.frame with five columns:

\itemize{
\item{term}{The term in the linear model being estimated and tested}
\item{estimate}{The estimated coefficient}
\item{stderror}{The standard error from the linear model}
\item{statistic}{t-statistic}
\item{p.value}{two-sided p-value}
}

These are the values contained in the coefficients matrix computed by \link{summary.lm}
(though with new column names)
}
\description{
tidy a linear model by returning a data frame version of the coefficients
table
}

