% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-lmrob-tidiers.R
\name{tidy.lmRob}
\alias{tidy.lmRob}
\alias{robust_tidiers}
\title{Tidy a(n) lmRob object}
\usage{
\method{tidy}{lmRob}(x, ...)
}
\arguments{
\item{x}{A \code{lmRob} object returned from \code{\link[robust:lmRob]{robust::lmRob()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{x}{An \code{lm} object created by \code{\link[stats:lm]{stats::lm()}}.}
  \item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}
  \item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}
  \item{quick}{Logical indiciating if the only the \code{term} and \code{estimate}
columns should be returned. Often useful to avoid time consuming
covariance calculations. Defaults to \code{FALSE}.}
  \item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}.
}
\examples{

library(robust)
m <- lmRob(mpg ~ wt, data = mtcars)

tidy(m)
augment(m)
glance(m)

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
glance(gm)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[robust:lmRob]{robust::lmRob()}}

Other robust tidiers: \code{\link{augment.glmRob}},
  \code{\link{augment.lmRob}}, \code{\link{glance.glmRob}},
  \code{\link{glance.lmRob}}, \code{\link{tidy.glmRob}}
}
\concept{robust tidiers}
