% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-lmrob-tidiers.R
\name{tidy.lmRob}
\alias{tidy.lmRob}
\alias{robust_tidiers}
\title{Tidy a(n) lmRob object}
\usage{
\method{tidy}{lmRob}(x, ...)
}
\arguments{
\item{x}{A \code{lmRob} object returned from \code{\link[robust:lmRob]{robust::lmRob()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
This method wraps \code{\link[=tidy.lm]{tidy.lm()}}.
}
\details{
For tidiers for robust models from the \pkg{MASS} package see
\code{\link[=tidy.rlm]{tidy.rlm()}}.
}
\examples{

library(robust)
m <- lmRob(mpg ~ wt, data = mtcars)

tidy(m)
augment(m)
glance(m)

gm <- glmRob(am ~ wt, data = mtcars, family = "binomial")
glance(gm)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[robust:lmRob]{robust::lmRob()}}

Other robust tidiers: \code{\link{augment.lmRob}},
  \code{\link{glance.glmRob}}, \code{\link{glance.lmRob}},
  \code{\link{tidy.glmRob}}
}
\concept{robust tidiers}
