% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tidy.NULL}
\alias{tidy.NULL}
\title{tidy on a NULL input}
\usage{
\method{tidy}{NULL}(x, ...)
}
\arguments{
\item{x}{A value NULL}

\item{...}{extra arguments (not used)}
}
\value{
An empty data.frame
}
\description{
tidy on a NULL input returns an empty data frame, which means it can be
combined with other data frames (treated as "empty")
}

