% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcutt-tidiers.R
\name{tidy.orcutt}
\alias{tidy.orcutt}
\alias{orcutt_tidiers}
\title{Tidy a(n) orcutt object}
\usage{
\method{tidy}{orcutt}(x, ...)
}
\arguments{
\item{x}{An \code{orcutt} object returned from \code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
This method wraps \code{\link[=tidy.lm]{tidy.lm()}}.
}
\examples{

library(orcutt)

reg <- lm(mpg ~ wt + qsec + disp, mtcars)
tidy(reg)

co <- cochrane.orcutt(reg)
co

tidy(co)
glance(co)

}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[orcutt:cochrane.orcutt]{orcutt::cochrane.orcutt()}}

Other orcutt tidiers: \code{\link{glance.orcutt}}
}
\concept{orcutt tidiers}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns:
  \item{estimate}{The estimated value of the regression term.}
  \item{p.value}{The two-sided p-value associated with the observed statistic.}
  \item{statistic}{The value of a T-statistic to use in a hypothesis that the regression term is non-zero.}
  \item{std.error}{The standard error of the regression term.}
  \item{term}{The name of the regression term.}

}
