% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy}
\alias{tidy}
\title{Turn a model object into a tidy tibble}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with information about model components.
}
\description{
Turn a model object into a tidy tibble
}
