% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-rlm-tidiers.R
\name{tidy.rlm}
\alias{tidy.rlm}
\title{Tidy a(n) rlm object}
\usage{
\method{tidy}{rlm}(x, ...)
}
\arguments{
\item{x}{An \code{rlm} object returned by \code{\link[MASS:rlm]{MASS::rlm()}}.}

\item{...}{Arguments passed on to \code{tidy.lm}
\describe{
  \item{}{}
}}
}
\value{
Something goes here
}
\description{
Description of what a \code{tidy} method does goes here. Wraps
\code{\link[=tidy.lm]{tidy.lm()}}. Details on this.  Document params other than \code{x}.
}
\details{
For tidiers for models from the \pkg{robust} package see
\code{\link[=tidy.lmRob]{tidy.lmRob()}} and \code{\link[=tidy.glmRob]{tidy.glmRob()}}.
}
\seealso{
\code{\link[=tidy]{tidy()}}, \code{\link[=tidy.lm]{tidy.lm()}}

\code{\link[MASS:rlm]{MASS::rlm()}}

Other rlm tidiers: \code{\link{augment.rlm}},
  \code{\link{glance.rlm}}
}
\concept{rlm tidiers}
