% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-tidiers.R
\name{tidy.table}
\alias{tidy.table}
\title{Tidy a(n) table object}
\usage{
\method{tidy}{table}(x, ...)
}
\arguments{
\item{x}{A \link{table} object.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} in long-form containing frequency information
for the table in a \code{Freq} column. The result is much like what you get
from \code{\link[tidyr:gather]{tidyr::gather()}}.
}
\description{
Tidy summarizes information about the components of a model.
A model component might be a single term in a regression, a single
hypothesis, a cluster, or a class. Exactly what tidy considers to be a
model component varies cross models but is usually self-evident.
If a model has several distinct types of components, you will need to
specify which components to return.
}
\details{
Directly calls \code{\link[tibble:as_tibble]{tibble::as_tibble()}} on a \link{table} object, which
does the same things as \code{\link[=as.data.frame.table]{as.data.frame.table()}} but also gives the
returned object \link[tibble:tibble]{tibble::tibble} class.
}
\examples{

tab <- with(airquality, table(cut(Temp, quantile(Temp)), Month))
tidy(tab)

}
\seealso{
\code{\link[=as_tibble.table]{as_tibble.table()}}
}
