% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats_tidiers.R
\name{tidy.ts}
\alias{tidy.ts}
\title{tidy a ts timeseries object}
\usage{
\method{tidy}{ts}(x, ...)
}
\arguments{
\item{x}{a "ts" object}

\item{...}{extra arguments (not used)}
}
\value{
a tidy data frame
}
\description{
Turn a ts object into a tidy data frame. Right now simply uses
\code{as.data.frame.ts}.
}
\seealso{
\link{as.data.frame.ts}
}

