% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-vector-tidiers.R
\name{tidy.numeric}
\alias{tidy.numeric}
\alias{tidy.character}
\alias{tidy.logical}
\title{Tidy atomic vectors}
\usage{
\method{tidy}{numeric}(x, ...)

\method{tidy}{character}(x, ...)

\method{tidy}{logical}(x, ...)
}
\arguments{
\item{x}{An object of class "numeric", "integer", "character", or "logical".
Most likely a named vector}

\item{...}{Extra arguments (not used)}
}
\description{
Vector tidiers are deprecated and will be removed from an upcoming release
of broom.
}
\details{
Turn atomic vectors into data frames, where the names of the vector (if they
exist) are a column and the values of the vector are a column.
}
\examples{

\dontrun{
x <- 1:5
names(x) <- letters[1:5]
tidy(x)
}

}
